% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-structure.R
\name{assignCellsToNodes}
\alias{assignCellsToNodes}
\title{Assign Cells to Nodes}
\usage{
assignCellsToNodes(object, node.size = 100, pseudotime)
}
\arguments{
\item{object}{An URD object}

\item{node.size}{(Numeric) Number of cells per node}

\item{pseudotime}{(Character) Pseudotime to use (i.e. a column name of \code{@pseudotime})}
}
\description{
Assign Cells to Nodes
}
