% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-exp.R
\name{aucprTestAlongTreeOld}
\alias{aucprTestAlongTreeOld}
\title{Test for differential gene expression along tree using binomial test}
\usage{
aucprTestAlongTreeOld(object, pseudotime, tips, log.effect.size = 0.25,
  min.auc = 0.1, frac.must.express = 0.1, frac.min.diff = 0.1,
  genes.use = NULL, root = NULL, only.return.global = F,
  must.beat.sibs = 0.5, report.stats = F)
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character) Name of pseudotime to use}

\item{tips}{(Character vector) The tip of the trajectory to find differential expression for (either as tip number or name)}

\item{log.effect.size}{(Numeric) Minimum fold-change difference in expression to be considered differential (default, \code{log(2)} is 2-fold change)}

\item{min.auc}{(Numeric) Minimum area under the curve for being considered differential}

\item{frac.must.express}{(Numeric) Fraction of cells of interest that must express the gene to consider it differential.}

\item{frac.min.diff}{(Numeric) Minimum difference in fraction of cells expressing a gene to consider it differential.}

\item{genes.use}{(Character vector) Which genes to test for differential expression (default \code{NULL} is all genes)}

\item{root}{(Character) Which segment of the tree (by number) to use as the root for differential express (default \code{NULL} will autodetect, but this can be useful if you do not want to traverse up the entire tree)}

\item{only.return.global}{(Logical) Genes must also pass an overall in-trajectory vs. out-of-trajectory test across all cells in the data to be considered differential (default, \code{FALSE})}

\item{must.beat.sibs}{(Numeric) At multi-furcated branchpoints, independent comparisons are made to each sibling. For a gene to be considered differential it must be a marker against this proportion of its sibling branches. (1 means it must be a marker against all other branches, 0 means it must be a marker against a single other branch.)}

\item{report.stats}{(Logical) Should information be returned about all of the comparisons? (nGene, nTrans, pseudotime, n.cells) If \code{TRUE}, this function returns a list instead of a data.frame}
}
\value{
a data.frame of gene expression results if \code{report.stats=F} or a list with entries \code{diff.exp} and \code{stats} if \code{report.stats=T}.
}
\description{
This test performs differential expression testing along URD's recovered tree
structure using a binomial expression test (see: \code{\link{markersBinom}}).
}
\details{
It starts from the segment provided in \code{tips} and compares it to each of
its siblings (and their descendants) matched to the same pseudotime limits as
the segment under consideration. Markers that are upregulated in a segment compared
to either any of its siblings (if \code{must.beat.all.sibs=FALSE}) or compared
to each of its siblings (if \code{must.beat.all.sibs=TRUE}) become putative
markers.
}
