% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-exp.R
\name{aucprTestByFactor}
\alias{aucprTestByFactor}
\title{Test for differential gene expression along tree using binomial test}
\usage{
aucprTestByFactor(
  object,
  cells.1,
  cells.2,
  label,
  groups,
  log.effect.size = 0.25,
  auc.factor = 1,
  min.auc.thresh = 0.1,
  max.auc.thresh = Inf,
  frac.must.express = 0.1,
  frac.min.diff = 0,
  genes.use = NULL,
  min.groups.to.mark = 1,
  report.debug = F
)
}
\arguments{
\item{object}{An URD object}

\item{cells.1}{(Character vector) Cells in population of interest}

\item{cells.2}{(List) Cells to compare against. If a list, each entry is a population to compare against, and markers must beat all populations.}

\item{label}{(Character) Label to use to split cells. Must be a column of \code{group.ids}}

\item{groups}{(List) List of label values to include in each group.}

\item{log.effect.size}{(Numeric) Minimum fold-change difference in mean expression to be considered differential (default, \code{log(2)} is 2-fold change)}

\item{auc.factor}{(Numeric) The precision-recall AUC is determined for a random classifier is determined
based on the size of populations. To be considered differential, genes must have an AUC this factor
multiplied by the expected AUC of a random classifier.}

\item{frac.must.express}{(Numeric) Fraction of cells of interest that must express the gene to consider it differential.}

\item{frac.min.diff}{(Numeric) Minimum difference in fraction of cells expressing a gene to consider it differential.}

\item{genes.use}{(Character vector) Which genes to test for differential expression (default \code{NULL} is all genes)}

\item{min.groups.to.mark}{(Numeric) For how many of the groups must a gene be a marker in order to be considered differential?}

\item{report.debug}{(Logical) If \code{TRUE}, this function returns a list instead of a data.frame, with 
\code{$stats} containing information about all of the comparisons? (nGene, nTrans, pseudotime, n.cells) and
\code{$marker.chain} containing the markers from each branchpoint.}

\item{min.auc.threshold}{(Numeric) This acts as a lower bound for the AUC threshold, no matter how unbalanced the populations are.}

\item{max.auc.threshold}{(Numeric) This acts as an upper bound for how high the AUC must be for a gene to be considered differential.}
}
\value{
a data.frame of gene expression results if \code{report.debug=F} or a list with entries \code{diff.exp} and \code{stats} if \code{report.debug=T}.
}
\description{
Test for differential gene expression along tree using binomial test
}
