% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-exp.R
\name{aucprThreshold}
\alias{aucprThreshold}
\title{Precision-Recall AUC threshold}
\usage{
aucprThreshold(cells.1, cells.2, factor = 1, max.auc = Inf)
}
\arguments{
\item{cells.1}{(Character vector) List of cells}

\item{cells.2}{(Character vector) List of cells}

\item{factor}{(Numeric) Factor to multiply AUC of random classifier by to
determine threshold}

\item{max.auc}{(Numeric) Upper limit of AUC threshold to return}
}
\value{
(Numeric) AUC threshold
}
\description{
The AUC for a random classifier for a precision-recall curve is based on the 
number of cells in each population. This function determines the random AUC,
multiplies it by an arbitrary factor (how much better than random must a
classifier be?), and gives it an upper cap.
}
