% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batchComBat}
\alias{batchComBat}
\title{Batch correct data using ComBat}
\usage{
batchComBat(
  object,
  batch.column,
  adjustment.columns = NULL,
  reference.batch = NULL,
  genes.correct = NULL,
  thresh.out = 0.05,
  min.out = 0,
  max.out = (ceiling(max(object@logupx.data)) + 1),
  parametric.prior = T,
  prior.plots = F,
  verbose = F
)
}
\arguments{
\item{object}{An URD object}

\item{batch.column}{(Character) Name of column in \code{@meta} that encodes batch information.}

\item{adjustment.columns}{(Character vector) Name of columns in \code{@meta} that encode information to make sure not to correct (e.g. developmental stage, etc.)}

\item{reference.batch}{(Character) Name of reference batch to hold constant (default \code{NULL} adjusts all batches toward each other)}

\item{genes.correct}{(Character vector) Which genes to correct (default \code{NULL} corrects all genes)}

\item{thresh.out}{(Numeric) To keep matrix sparse, change values less than this to \code{min.out}}

\item{min.out}{(Numeric) Minimum gene expression value to return}

\item{max.out}{(Numeric) Maximum gene expression value to return (Default: capped at maximum gene expression value of original data + 1.)}

\item{parametric.prior}{(Logical) ComBat option: Assume normal distributed gene expression (Warning: extremely slow if FALSE)}

\item{prior.plots}{(Logical) ComBat option: Plots to compare empirical and parametric batch (see \code{\link[sva]{ComBat}} for more info.)}

\item{verbose}{(Logical) Print progress}
}
\value{
An URD object with logupx.data batch-corrected using ComBat.
}
\description{
Batch correct data using ComBat
}
