% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchpoint-investigation.R
\name{branchpointDetailsVisitDist}
\alias{branchpointDetailsVisitDist}
\title{Investigate Branchpoint - Visit Frequency Distributions}
\usage{
branchpointDetailsVisitDist(object, seg.1, seg.2, file = NULL,
  file.width = 375, file.height = 300)
}
\arguments{
\item{object}{An URD object}

\item{seg.1}{(Character) Segment 1 of fusion}

\item{seg.2}{(Character) Segment 2 of fusion}

\item{file}{(Character) Path to save plot to. (If \code{NULL}, will display live)}

\item{file.width}{(Numeric) Width per plot (in pixels) if saving file.}

\item{file.height}{(Numeric) Height per plot (in pixels) if saving file.}
}
\value{
Nothing. If \code{file==NULL}, will display a plot, otherwise will produce a file.
}
\description{
For a given pair of segments, this shows the branchpoint fusion details in the form of a series of distributionplots.
Each moving window through pseudotime that was considered is rendered as a plot, in order of pseudotime. (So,
the first plots are near the root.) The distribution of visitation frequency from the two segments is plotted in
red (seg.1) and green (seg.2). Plots are titled with the window number, p-value (drawn from the test used during
construction of the tree), and the determination of whether cell visitation within the window is significantly
different (TRUE) or not (FALSE).
}
