% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchpoint-investigation.R
\name{branchpointDetailsVisitTsne}
\alias{branchpointDetailsVisitTsne}
\title{Investigate Branchpoint - tSNE Visitation Groups}
\usage{
branchpointDetailsVisitTsne(
  object,
  seg.1,
  seg.2,
  file = NULL,
  file.width = 750,
  file.height = 600
)
}
\arguments{
\item{object}{An URD object}

\item{seg.1}{(Character) Segment 1 of fusion}

\item{seg.2}{(Character) Segment 2 of fusion}

\item{file}{(Character) Path to save plot to. (If \code{NULL}, will display live)}

\item{file.width}{(Numeric) Width per plot (in pixels) if saving file.}

\item{file.height}{(Numeric) Height per plot (in pixels) if saving file.}
}
\value{
Nothing. If \code{file==NULL}, will display a plot, otherwise will produce a file.
}
\description{
For a given pair of segments, this shows the branchpoint fusion details in the form of a series of tSNE plots.
Each moving window through pseudotime that was considered is rendered as a plot, in order of pseudotime. (So,
the first plots are near the root.) Cells within the window are
colored according to their visitation frequency from the two segments in a dual color plot. Red is visitation
from one segment, green is visitation from the other segment. Cells that are visited equally will be yellow.
}
