% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchpointPlot.R
\name{branchpointPreferenceLayout}
\alias{branchpointPreferenceLayout}
\title{Branchpoint Preference Layout}
\usage{
branchpointPreferenceLayout(object, pseudotime, lineages.1, lineages.2,
  parent.of.lineages, opposite.parent, min.visit = 0)
}
\arguments{
\item{object}{An URD object}

\item{lineages.1}{(Character) Segment(s) for the left side of the branchpoint}

\item{lineages.2}{(Character) Segment(s) for the right side of the branchpoint}

\item{parent.of.lineages}{(Character) Segment that is upstream of the branchpoint}

\item{opposite.parent}{(Character vector) Siblings of the \code{parent.of.lineages}.}

\item{min.visit}{(Numeric) Minimum visitation frequency to appear in the layout}
}
\value{
A branchpoint layout that can be used as input to \code{\link{plotBranchpointByGene}}
}
\description{
Branchpoint Preference Layout
}
