% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{calcKNN}
\alias{calcKNN}
\title{Calculate k-nearest neighbor graph}
\usage{
calcKNN(object, genes.use = object@var.genes, nn = 100)
}
\arguments{
\item{object}{An URD object}

\item{genes.use}{Genes to use for distance (default is variable genes, NULL is all genes.)}

\item{nn}{(Numeric) Number of nearest neighbors}
}
\value{
An URD object with k-nn graph in slot \code{knn}.
}
\description{
Calculates and stores a k-nearest neighbor graph based on Euclidean distance in 
gene expression space. This is useful for finding and removing outliers prior
to calculating a diffusion map.
}
