% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne.R
\name{calcTsne}
\alias{calcTsne}
\title{Calculate tSNE projection of data}
\usage{
calcTsne(object, dim.use = c("pca", "dm"),
  which.dims = which(object@pca.sig), perplexity = 30, theta = 0.5,
  max_iter = 1000, verbose = FALSE)
}
\arguments{
\item{object}{An URD object}

\item{dim.use}{(Character) Whether to calculate the tSNE projection from PCs (\code{"pca"}) or diffusion components (\code{"dm"}). Default is from PCs.}

\item{which.dims}{(Numeric vector) Which PCs (or diffusion components) to use. Defaults to the significant PCs. (The default will probably work with diffusion components, though it is non-sensical in that case.)}

\item{theta}{(Numeric) Speed/accuracy trade-off for Barnes-Hut approximation of tSNE (0-1, 0 is exact tSNE, higher is less accurate, default is 0.5)}

\item{max_iter}{(Numeric) Number of iterations to perform}

\item{verbose}{(Logical) Should Rtsne print progress updates?}

\item{perlexity}{(Numeric) Perplexity parameter for the tSNE}
}
\value{
An URD object with tSNE coordinates stored in the \code{tsne.y} slot.
}
\description{
Calculate tSNE projection of data
}
