% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood.R
\name{cellDistByExpression}
\alias{cellDistByExpression}
\title{Calculates the distance matrix between cells given their gene expression}
\usage{
cellDistByExpression(
  object,
  genes.use = object@var.genes,
  return.as.matrix = T
)
}
\arguments{
\item{object}{An URD object}

\item{genes.use}{(Character vector) Which genes to use for distance calculation (default: variable genes. NULL is all genes)}

\item{return.as.matrix}{(Logical) Return value as a matrix (memory hungry, but used for other URD functions) or as a dist object.}
}
\value{
Returns the distance between cells in gene expression space, either as a matrix
(if \code{return.as.matrix=T}) or a dist object.
}
\description{
Calculates distance between cells in gene expression space. Can be passed to \code{\link{tipPotential}}.
}
