% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-navigate.R
\name{cellsAlongLineage}
\alias{cellsAlongLineage}
\title{Extract cells along a lineage pathway}
\usage{
cellsAlongLineage(object, segments, remove.root = T)
}
\arguments{
\item{object}{An URD object}

\item{segments}{(Character vector) Terminal segment numbers or names}

\item{remove.root}{(Logical)}
}
\value{
A character vector of cell names
}
\description{
Extract cells along a lineage pathway
}
