% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cellsInCluster}
\alias{cellsInCluster}
\title{Return names of cells in a given cluster}
\usage{
cellsInCluster(object, clustering, cluster)
}
\arguments{
\item{object}{An URD object}

\item{clustering}{(Character) Name of a column in \code{@group.ids}}

\item{cluster}{(Character vector) Name of 1 or more clusters to return.}
}
\value{
Character vector of cell names
}
\description{
Return names of cells in a given cluster
}
\examples{
# Get members of cluster 15 from Louvain-10 clustering
cellsInCluster(object, clustering="Louvain=10", cluster="15")

# Get members of segments 29, 32, and 79 from reconstructed tree
cellsInCluster(object, clustering="segment", cluster=c("29","32","79"))

}
