% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterCentroids}
\alias{clusterCentroids}
\title{Cluster centroids in gene expression space}
\usage{
clusterCentroids(object, clustering, genes.use = object@var.genes)
}
\arguments{
\item{object}{An URD object}

\item{clustering}{(Character) Name of a clustering (i.e. a column in \code{@group.ids})}

\item{genes.use}{(Character vector) Genes to use for calculating distance (default: variable genes, NULL is all genes)}
}
\value{
A data.frame with mean gene expression per cluster in rows, and clusters as columns.
}
\description{
Cluster centroids in gene expression space
}
