% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood.R
\name{clusterTipPotential}
\alias{clusterTipPotential}
\title{Tip Potential of Clusters}
\usage{
clusterTipPotential(object, pseudotime, clustering,
  name.store = "tip.potential")
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character) Pseudotime to use (i.e. a column of \code{@pseudotime})}

\item{clustering}{(Character) Clustering to use (i.e. a column of \code{@group.ids})}

\item{name.store}{(Character) Name to store this tip potential calculation under (a column of \code{@pseudotime})}
}
\value{
An URD object with a column in \code{@pseudotime} named according to \code{name.store}.
}
\description{
Calculate likelihood that each cluster is a potential tip.
}
\details{
This assumes that clusters that are tips will be primarily connected to cells with younger pseudotime.
Thus, this function calculates the pseudotime difference between each cell in the data and mean pseudotime
of each cluster. Then, the pseudotime differences are averaged To do this,
the cumulative transition probability from all cells in the cluster to all cells outside of the clusters
is calculated. Then, the pseudotime difference between each cell and the mean of the cluster is calculated,
and a weighted average is performed according to the cumulative transition probability.
}
