% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-structure.R
\name{collapseShortSegments}
\alias{collapseShortSegments}
\title{Collapse Short Segments}
\usage{
collapseShortSegments(
  object,
  min.cells.per.segment = 1,
  min.pseudotime.per.segment = 0.01,
  collapse.root = F
)
}
\arguments{
\item{object}{An URD object}

\item{min.cells.per.segment}{(Numeric) Remove segments with fewer cells assigned}

\item{min.pseudotime.per.segment}{(Numeric) Remove segments shorter than this in pseudotime}

\item{collapse.root}{(Logical) Is the very root most segment OK to delete? (Sometimes it would have no segments, but be important not to collapse)}
}
\description{
This refines the dendrogram structure found by URD by removing overly short segments.
It is automatically called by \code{\link{buildTree}}, but can be re-run if needed.
}
\keyword{internal}
