% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood.R
\name{combine.probs}
\alias{combine.probs}
\title{Combine probabilities}
\usage{
combine.probs(x)
}
\arguments{
\item{x}{(Numeric Vector) Probabilities}
}
\value{
(Numeric) Cumulative probability
}
\description{
Calculates cumulative probability
}
