% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-smooth.R
\name{combineSmoothFit}
\alias{combineSmoothFit}
\title{Gene Smooth: Combine smoothed fits}
\usage{
combineSmoothFit(smoothed.fit, fits.use)
}
\arguments{
\item{smoothed.fit}{(List) A list of smoothed.fits to combine}

\item{fits.use}{(Character) Names of smoothed.fits in the list to use}
}
\value{
(List)
}
\description{
???
}
\keyword{internal}
