% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-tips.R
\name{combineTipVisitation}
\alias{combineTipVisitation}
\title{Combine Tip Visitation}
\usage{
combineTipVisitation(object, tip.1, tip.2, new.tip)
}
\arguments{
\item{object}{An URD object}

\item{tip.1}{(Character) Name of a tip to combine}

\item{tip.2}{(Character) Name of another tip to combine}

\item{new.tip}{(Character) Name to call the combined tip}
}
\value{
An URD object
}
\description{
If two tips were used to do random walks, but then seem to be the exact same population,
the random walks can be combined (biased by the number of cells in each tip), to combine
their visitation frequencies. Requires \code{\link{loadTipCells}} to be run prior.
}
