% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corner}
\alias{corner}
\title{Corner of data frame or matrix}
\usage{
corner(data.frame, n = 5)
}
\arguments{
\item{data.frame}{Input data.frame or matrix}

\item{n}{(Numeric) Number of rows and columns to return (default: 5)}
}
\description{
Returns the upper-left corner of a data.frame or matrix
}
