% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urd-class.R
\name{createURD}
\alias{createURD}
\title{Create a new URD object}
\usage{
createURD(
  count.data,
  meta = NULL,
  min.cells = 3,
  min.genes = 500,
  min.counts = 10,
  gene.max.cut = 5000,
  max.genes.in.ram = 5000,
  verbose = T
)
}
\arguments{
\item{count.data}{(Matrix or dgCMatrix) UMI expression data, with rows as genes and columns as cells}

\item{meta}{(data.frame) Metadata, with rows as cells (row names should match column names of \code{count.data})}

\item{min.cells}{(Numeric) Minimum number of cells that must express a gene to retain it}

\item{min.genes}{(Numeric) Minimum number of genes a cell must express to retain it}

\item{min.counts}{(Numeric) Minimum number of UMIs detected for a gene across the entire data to retain it}

\item{gene.max.cut}{(Numeric) Maximum number of UMIs observed for a gene in a single cell to retain it. If \code{NA}, this step will be skipped.}

\item{max.genes.in.ram}{(Numeric) Number of genes to normalize and log-transform at a time (to prevent running out of memory as matrices become non-sparse during the process)}
}
\value{
An URD object
}
\description{
Creates a new URD object. Provide expression data as UMI counts and (optionally)
additional metdata. This function will perform basic filtering of the data
(though the user may want to perform more advanced filtering before creating
the URD object, using \code{ds.meta.trim}), will normalize and log2 transform the
data, and will return an URD object with the original data in slot \code{count.data},
the normalized log-transformed data in slot \code{logupx.data}, the metadata in slot
\code{meta}, and an initial grouping of the data drawn from the cell names, up to the
first dash (-) or underscore (_) in slot \code{group.ids}.
}
