% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-smooth.R
\name{cropSmoothFit}
\alias{cropSmoothFit}
\title{Gene Smooth: Crop to particular pseudotimes}
\usage{
cropSmoothFit(smoothed.fit, pt.min = -Inf, pt.max = Inf)
}
\arguments{
\item{smoothed.fit}{(List) Result from \code{\link{geneSmoothFit}}}

\item{pt.min}{(Numeric) Minimum pseudotime to retain}

\item{pt.max}{(Numeric) Maximum pseudotime to retain}
}
\value{
(List) Result from \code{\link{geneSmoothFit}}, but cropped to the range of \code{pt.min} and \code{pt.max}.
}
\description{
This crops a result from \code{\link{geneSmoothFit}} to a particular pseudotime range. Can be used to split a smoothed result into segments for labeling in \code{\link{plotSmoothFitMultiCascade}}.
}
