% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutstring}
\alias{cutstring}
\title{Cut string into fields}
\usage{
cutstring(x, delimiter, field)
}
\arguments{
\item{x}{(Character) A character vector}

\item{delimiter}{(Character) Delimiter character to use to split}

\item{field}{(Numeric) Which field to return; count starts at left with 1.}
}
\value{
Character vector
}
\description{
Reworks \code{\link{strsplit}} to act like Unix cut. Instead of returning a list, returns a vector of the chosen field.
}
