% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-for-plot.R
\name{defaultURDContinuousColors}
\alias{defaultURDContinuousColors}
\title{Default URD continuous colors}
\usage{
defaultURDContinuousColors(with.grey = F, symmetric = F)
}
\value{
Vector of default colors
}
\description{
Default URD continuous colors
}
\keyword{internal}
