% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-exp.R
\name{differentialAUCPR}
\alias{differentialAUCPR}
\title{Area under precision-recall curve for determining marker specificity.}
\usage{
differentialAUCPR(x, y)
}
\arguments{
\item{x}{(Numeric vector) Gene expression values from one population}

\item{y}{(Numeric vector) Gene expression values from the other population}
}
\value{
(Numeric) Area under the precision-recall curve
}
\description{
Area under precision-recall curve for determining marker specificity.
}
\keyword{internal}
