% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-structure.R
\name{divergenceKSVisitation}
\alias{divergenceKSVisitation}
\title{Calculate visitation divergence using KS-test}
\usage{
divergenceKSVisitation(
  visit.data,
  pseudotime.windows,
  cells.segment.1,
  cells.segment.2
)
}
\arguments{
\item{visit.data}{(data.frame) Rows are cells, columns are: "segment.1" and "segment.2" (visitation frequency from random walks
initiated in segment 1 and 2), "pseudotime" (the pseudotime of each cell), and "pseudotime.group" (the number of the pseudotime
bin each cell belongs to)}

\item{pseudotime.windows}{(matrix) Rows are pseudotime windows, and the columns in each row represent the pseudotime bins that belong to that window.}

\item{cells.segment.1}{(Character vector) Cells in segment 1}

\item{cells.segment.2}{(Character vector) Cells in segment 2}
}
\value{
(data.frame) Rows are pseudotime windows, columns are KS-test p-value ("p"), pseudotime of cells in the window ("mean.pseudotime", "min.pseudotime", "max.pseudotime"), and number of cells considered from each segment ("cells.visited.seg1", "cells.visited.seg2")
}
\description{
Calculate visitation divergence using KS-test
}
\keyword{internal}
