% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{dsCombineDGE}
\alias{dsCombineDGE}
\title{Fuse Dropseq Digital Gene Expression tables}
\usage{
dsCombineDGE(list.dge)
}
\arguments{
\item{list.dge}{(List) List of DGEs}
}
\value{
Sparse matrix (dgCMatrix) with rows as genes and columns as cells.
}
\description{
Combines multiple DGEs with different dimensions, keeping all genes present
in any DGE, and correctly setting values to 0 if that gene was not observed
in another sample.
}
\seealso{
\code{\link{ds.read.dge}} to read in DGEs to provide in \code{list.dge}.
}
