% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{dsMetaTrim}
\alias{dsMetaTrim}
\title{Plot and trim Dropseq DGE tables by metadata}
\usage{
dsMetaTrim(dge.table, ds.meta, cell.name = "CELL", meta.name,
  meta.min = NULL, meta.max = NULL, title = "UMIs per cell",
  x.lim = NULL, breaks = 150)
}
\arguments{
\item{dge.table}{(Matrix or data.frame) Gene expression table}

\item{ds.meta}{(data.frame) Metadata}

\item{cell.name}{(Character) Column name of \code{ds.meta} that corresponds to column names of \code{dge.table} (default: "CELL")}

\item{meta.name}{(Character) Column name of \code{ds.meta} to use for filtering cells}

\item{meta.min}{(Numeric) Minimum acceptable value of metadata (default: NULL)}

\item{meta.max}{(Numeric) Maximum acceptable value of metadata (default: NULL)}

\item{title}{(Character) Histogram title (default: "UMIs per cell")}

\item{breaks}{(Numeric) Histogram bins (default: 150)}

\item{xlim}{(Numeric) Histogram x-axis limits (default: NULL)}
}
\description{
Plot and trim Dropseq DGE tables by metadata
}
