% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{dsReadDGE}
\alias{dsReadDGE}
\title{Read Dropseq Digital Gene Expression tables}
\usage{
dsReadDGE(dge.file, dge.name = NULL)
}
\arguments{
\item{dge.file}{(Character) Path to digital gene expression table}

\item{dge.name}{(Character) Name to append to all cell barcodes if multiple tables will be combined (default: NULL)}
}
\value{
DGE: a sparse Matrix (dgCMatrix) with rows of genes and columns of cells.
}
\description{
Reads digital gene expression tables output by the Dropseq pipeline.
}
