% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{dsReadStats}
\alias{dsReadStats}
\title{Read cell barcode statistics produced by Dropseq pipeline}
\usage{
dsReadStats(dge.summary.reports, cell.readcount.reports = NULL, sample.names,
  name.cells.with.sample = T)
}
\arguments{
\item{dge.summary.reports}{(Character vector) Paths of DGE report files. (These files typically end with the suffix ".report_dge.summary.txt")}

\item{cell.readcount.reports}{(Character vector) Paths of cell readcount reports for getting number of reads per cell.  (These files typically end with the suffix ".report_cell_readcounts.txt.gz")}

\item{sample.names}{(Character vector) Names of samples.}

\item{name.cells.with.sample}{(Logical) Convert cell barcodes to cell names?}
}
\value{
Returns a data.frame with cells as rows, and metadata as columns.
}
\description{
The Dropseq pipeline produces output summary files that describe the number of
genes, UMIs, and reads per cell barcode. This reads them in from several files
and collates them into a single data.table.
}
