% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{fdlDensity}
\alias{fdlDensity}
\title{Calculate force-directed layout local density}
\usage{
fdlDensity(object, neighbor = 10, knn = NULL)
}
\arguments{
\item{object}{An URD object}

\item{neighbor}{(Numeric) Distance to \code{neighbor}th nearest neighbor is used for
density adjustment of point transparency on force-directed layout plots.}
}
\description{
Add distance to nth nearest neighbor to force directed layout so that you can later
calculate local density and adjust alpha values.
}
