% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var-genes.R
\name{findVariableGenes}
\alias{findVariableGenes}
\title{Find variable genes}
\usage{
findVariableGenes(object, cells.fit = NULL, set.object.var.genes = T,
  diffCV.cutoff = 0.5, mean.min = 0.005, mean.max = 100, main.use = "",
  do.plot = T)
}
\arguments{
\item{object}{(URD object): Uses expression data from \code{object@count.data[,cells.fit]}}

\item{cells.fit}{(Character Vector): Which cells to use for finding variable genes (if \code{NULL}, uses all cells.)}

\item{set.object.var.genes}{(Logical): Return an object with \code{var.genes} set (if \code{TRUE}) or return a character vector of variable genes (if \code{FALSE})}

\item{diffCV.cutoff}{(Numeric): Difference in coefficient of variation (CV) between null distribution & genes that will be considered variable (Difference is in log space, so this amounts to a fold-change)}

\item{mean.min}{(Numeric): Genes must have this minimum mean expression to be selected (Use to eliminate noisy lowly expressed genes)}

\item{mean.max}{(Numeric): Genes must have less than this maximum mean expression to be selected (Use to eliminate the high end if the null distribution fits very poorly there)}

\item{main.use}{(Character): title to display for the overall three-panel plot}

\item{do.plot}{(Logical): Display plots}
}
\value{
Either an URD object with \code{var.genes} set (if \code{set.object.var.genes=T}) or character vector of variable genes (if \code{set.object.var.genes=F})
}
\description{
Find variable genes
}
