% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood.R
\name{floodBuildTM}
\alias{floodBuildTM}
\title{Build pseudotime transition matrix}
\usage{
floodBuildTM(object = NULL, dm = NULL)
}
\arguments{
\item{object}{An URD object}

\item{dm}{A Diffusion Map}
}
\value{
A sparse matrix (dgCMatrix) of normalized transition probabilities.
}
\description{
This scales the transition probabilities so that maximum sum of transition
probabilities for a cell is 1. This modified matrix is used for calculating pseudotime. It requires
either the user to provide an URD object (\code{object=}) with a diffusion map
stored in slot \code{@dm}, or a diffusion map directly (\code{dm=}).
}
\details{
This function can be called directly and the transition matrix stored, or
\code{\link{floodPseudotime}} will call this function automatically if
one isn't provided.

Thanks to @diegomscoelho for suggested improvements.
}
