% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood.R
\name{floodPseudotime}
\alias{floodPseudotime}
\title{Calculate pseudotime by 'flooding'}
\usage{
floodPseudotime(object, root.cells, n = 20, minimum.cells.flooded = 2,
  tm.flood = NULL, verbose = F, verbose.freq = 10)
}
\arguments{
\item{object}{An URD object}

\item{root.cells}{(Character vector) Names of cells that will be considered the root and assigned pseudotime 0.}

\item{n}{(Numeric) Number of simulations to perform and average}

\item{minimum.cells.flooded}{(Numeric) Stop pseudotime calculation when fewer cells are newly visited in a given iteration, and assign remaining unvisited cells pseudotime NA. This is designed to prevent a few poor data points from throwing off the entire pseudotime calculation.}

\item{tm.flood}{(Sparse matrix) A sparse matrix of normalized transition probabilities (as returned by \code{\link{floodBuildTM}}. If unprovided (\code{NULL}), this will be calculated automatically.}

\item{verbose}{(Logical) Report on progress?}

\item{verbose.freq}{(Numeric) Give a report every this many iterations.}
}
\value{
(data.frame) Rows are cells, columns are flood simulations, values are the iteration that newly visited each cell. This can be processed into pseudotime and stored in an URD object with \code{\link{floodPseudotimeProcess}}.
}
\description{
This calculates pseudotime by performing a probabilistic breadth-first search
of the k-nearest neighbor graph that was used to generate the diffusion map
calculated on the data.
}
\details{
A group of 'root' cells are pre-initialized as visited, an an iterative search
of the graph is performed. The probability that a cell is newly visited is the
cumulative transition probability from all cells that have previously been
visited in the graph. The process continues until either the entire graph is
visited, or fewer than \code{minimum.cells.flooded} new cells are visited in
a given iteration.
}
