% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood.R
\name{floodPseudotimeCalc}
\alias{floodPseudotimeCalc}
\title{A single iteration of flood pseudotime calculation.}
\usage{
floodPseudotimeCalc(object, start.cells, minimum.cells.flooded = 2,
  tm.flood = NULL, verbose.freq = 0)
}
\arguments{
\item{object}{An URD object}

\item{start.cells}{(Character vector) Cells to initialize as visited to start pseudotime calculation}

\item{minimum.cells.flooded}{(Numeric) Stop pseudotime calculation when fewer cells are newly visited in a given iteration, and assign remaining unvisited cells pseudotime NA.}

\item{tm.flood}{(Sparse or full matrix) A sparse matrix of normalized transition probabilities. If unprovided (\code{NULL}), this will be calculated automatically.}

\item{verbose.freq}{(Numeric) Report progress after this many iterations. 0 is silent.}
}
\value{
(Numeric vector) The iteration that newly visited each cell.
}
\description{
This runs a single iteration of flood pseudotime. This function is meant to 
be called by \code{\link{floodPseudotime}} to run many iterations.
}
\keyword{internal}
