% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-cascade.R
\name{geneCascadeHeatmap}
\alias{geneCascadeHeatmap}
\title{Plot Gene Cascade Heatmap}
\usage{
geneCascadeHeatmap(cascade, color.scale = brewer.pal(9, "YlOrRd"),
  add.time = NULL, times.annotate = seq(0, 1, 0.1), title = "",
  annotation.list = NULL)
}
\arguments{
\item{cascade}{(list) A processed gene cascade from \code{\link{geneCascadeProcess}}}

\item{color.scale}{(Character vector)}

\item{add.time}{(Character or NULL) Either the name of a column of \code{@meta} that contains actual time information
to label the x-axis of the heatmap or NULL to instead use pseudotime}

\item{times.annotate}{(Numeric vector)}

\item{annotation.list}{}
}
\description{
Plot Gene Cascade Heatmap
}
