% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-cascade.R
\name{geneCascadeImpulsePlots}
\alias{geneCascadeImpulsePlots}
\title{Plot Impulse Fits For a Gene Cascade}
\usage{
geneCascadeImpulsePlots(cascade, file = NULL, verbose = F)
}
\arguments{
\item{cascade}{(List) A gene cascade, such as generated by \code{\link{geneCascadeProcess}}}

\item{file}{(Path) A path to save the plot to as PDF (if NULL, display)}
}
\description{
This allows checking that the impulse fits are working correctly. The
windowed data is plotted as black points for each gene, and the fit
is plotted in a thick line, color dependent on the type of fit that
was chosen: blue (linear), green (single sigmoid), red (double sigmoid).
Additionally, in thinner lines, other parameters are plotted. The onset
and offset times are plotted in blue (light for onset, dark for offset).
For sigmoid fits, the
}
