% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-cascade.R
\name{geneCascadeImpulsePlotsOld}
\alias{geneCascadeImpulsePlotsOld}
\title{Plot Impulse Fits For a Gene Cascade}
\usage{
geneCascadeImpulsePlotsOld(cascade, file = NULL, verbose = F)
}
\arguments{
\item{cascade}{(List) A gene cascade, such as generated by \code{\link{geneCascadeProcess}}}

\item{file}{(Path) A path to save the plot to as PDF (if NULL, display)}
}
\description{
This allows checking that the impulse fits are working correctly
}
