% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-cascade.R
\name{geneCascadeProcess}
\alias{geneCascadeProcess}
\title{Process a gene cascade}
\usage{
geneCascadeProcess(object, pseudotime, cells, genes,
  background.genes = sample(setdiff(rownames(object@logupx.data),
  object@var.genes), 1000), moving.window = 3, cells.per.window = 10,
  plot.scaled = T, verbose = T, verbose.genes = F,
  limit.single.sigmoid.slopes = c("none", "on", "off"), k = 50, a = 0.05)
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character) Name of pseudotime}

\item{cells}{(Character vector) Cells to include}

\item{genes}{(Character vector) Genes to include}

\item{background.genes}{(Character vector) Genes to use for calculation of background noise model}

\item{moving.window}{(Numeric)}

\item{cells.per.window}{(Numeric)}

\item{plot.scaled}{(Logical)}

\item{verbose}{(Logical)}

\item{verbose.genes}{(Logical) If \code{verbose=T}, should the start time of each gene be printed?}

\item{gene.cuts}{(Numeric) Number of groups}

\item{expression.thresh}{(Numeric)}

\item{ignore.expression.before}{(Numeric)}

\item{limit.single.sigmoid}{("none", "on", "off")}
}
\value{
List
}
\description{
Process a gene cascade
}
