% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-smooth.R
\name{geneSmoothFit}
\alias{geneSmoothFit}
\title{Gene Smooth: Fit expression of genes}
\usage{
geneSmoothFit(object, pseudotime, cells, genes, method = c("lowess",
  "spline"), moving.window = 3, cells.per.window = NULL,
  pseudotime.per.window = NULL, verbose = T, verbose.genes = F, ...)
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character) Name of pseudotime (i.e. a column name of \code{@pseudotime})}

\item{cells}{(Character vector) Cells to include}

\item{genes}{(Character vector) Genes to include}

\item{method}{(Character) Which smoothing method to use ("lowess" calls \code{\link{stats::lowess}} and "spline" calls \code{\link{stats::smooth.spline}})}

\item{moving.window}{(Numeric) Number of bins to use per window}

\item{cells.per.window}{(Numeric or \code{NULL}) Size of bins (number of cells)}

\item{pseudotime.per.window}{(Numeric or \code{NULL}) Size of bins (pseudotime)}

\item{verbose}{(Logical) Display progress reports?}

\item{verbose.genes}{(Logical) Display a message when starting each gene?}
}
\value{
(Named List):
\itemize{ 
  \item{\strong{\code{pt.windows}}: (List) Character vectors of cell IDs of cells in each window, named by either the mean, min, or max pseudotime of those cells (depending on \code{name.by})}
  \item{\strong{\code{mean.expression}}: (data.frame) Mean expression of each gene in each pseudotime window}
  \item{\strong{\code{scaled.expression}}: (data.frame) Mean expression of each gene in each pseudotime window, scaled to the maximum observed expression}
  \item{\strong{\code{mean.smooth}}: (data.frame) Value of fit curve for each pseudotime window (i.e. to predict value of mean.expression)}
  \item{\strong{\code{scaled.smooth}}: (data.frame) Value of fit curve for each pseudotime window, scaled to max expression (i.e. to predict value of scaled.expression)}
  \item{\strong{\code{mean.expression.red}}: (data.frame) Mean expression reduced to the same dimensions of \code{mean.smooth}. (When method is 'spline' if two windows have the same pseudotime, only a single prediction is made for that pseudotime, so this will average the two input windows.)}
  \item{\strong{\code{scaled.expression.red}}: (data.frame) Scaled expression reduced to the same dimensions of \code{scaled.smooth}.}
  \item{\strong{\code{method}}: (Character) Identify fit method ("impulse")}
}
}
\description{
This takes a group of genes and cells, averages gene expression in groups of cells (determined using a moving window through pseudotime), and then uses smoothing algorithms (e.g. LOESS or spline fitting) to describe the expression of each gene. The results are returned for use in \code{\link{plotSmoothFit}} to visualize the fits or several sets of results (e.g. from different trajectories) can be combined with \code{\link{cropSmoothFit}} and plotted with \code{\link{plotSmoothFitMultiCascade}}.
}
\examples{
# Fit a gene expression curve using "spline" method for
# genes "endo.genes" and cells in segments 47 and 46 in the tree.
tentacle.spline <- geneSmoothFit(urd.object, method = "spline", 
  pseudotime = "pseudotime", 
  cells = cellsInCluster(urd.object, "segment", c("47", "46")), 
  genes = endo.genes, moving.window = 1, cells.per.window = 5, spar = 0.9
)

}
