% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-smooth.R
\name{geneSmoothReduce}
\alias{geneSmoothReduce}
\title{Gene Smooth: Reduce dimensions of Smooth Fit}
\usage{
geneSmoothReduce(smoothed.fit, agg.func = mean)
}
\arguments{
\item{smoothed.fit}{(List) Output from \code{\link{geneSmoothFit}}}

\item{agg.func}{(function) Function to use to aggregate columns with same
pseudotime. Default is mean.}
}
\value{
(List) \code{smoothed.fit} input with \code{$mean.expression.red} and \code{$scaled.expression.red} added.
}
\description{
When multiple windows have the same pseudotime, and \code{\link{geneSmoothFit}}
is run with \code{method="spline"}, a single output point will be created,
creating smoothed expression data with different dimensions from the original
expression data. This function produces 'reduced' data frames of the same
dimensions and loads them into the smoothed.fit object by aggregating windows
with the same pseudotime.
}
\keyword{internal}
