% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{get.binary.data}
\alias{get.binary.data}
\title{Get binary data}
\usage{
get.binary.data(object, genes = NULL)
}
\arguments{
\item{object}{An URD object}

\item{genes}{(Character vector) Genes to return in binarized matrix (default, all genes)}
}
\value{
Sparse matrix (dgCMatrix)
}
\description{
Returns a sparse matrix of 0 and 1 showing whether any expression of each gene
was observed in each cell.
}
