% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{get.upx.data}
\alias{get.upx.data}
\title{Get UPX (un-logged) data}
\usage{
get.upx.data(object, genes = NULL)
}
\arguments{
\item{object}{An URD object}

\item{genes}{(Character vector) Genes to return in UPX atrix (default, all genes)}
}
\value{
Sparse matrix (dgCMatrix)
}
\description{
Returns a sparse matrix of expression values that are normalized for library
size (UPX normalization), but are not log transformed.
}
