% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{get.z.data}
\alias{get.z.data}
\title{Get z-scored data}
\usage{
get.z.data(object, genes = NULL)
}
\arguments{
\item{object}{An URD object}

\item{genes}{(Character vector) Genes to return in z-scored matrix (default, all genes)}
}
\value{
Matrix of z-scored data (Warning: memory hog as data will no longer be sparse.)
}
\description{
Returns z-scored data, calculated on the log-transformed, normalized data. Eliminates
any genes that produce NA values, which can occur if they have standard deviation 0.
}
