% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gridArrangeMulti}
\alias{gridArrangeMulti}
\title{Arrange multiple ggplots into several gridded pages}
\usage{
gridArrangeMulti(grobs, ncol = 2, nrow = 2)
}
\arguments{
\item{grobs}{(List) List of grobs (grid objects, e.g. returned by ggplot functions)}

\item{ncol}{(Numeric) Number of columns in each plot (default: 2)}

\item{nrow}{(Numeric) Number of rows in each plot (default: 2)}
}
\description{
Arranges several plots into a multi-page gridded graph
}
