% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cells.R
\name{groupFromCells}
\alias{groupFromCells}
\title{Create a group.id from a list of cells}
\usage{
groupFromCells(object, group.id, cells)
}
\arguments{
\item{object}{An URD object}

\item{group.id}{(Character) Name of the \code{group.id} to store.}

\item{cells}{(Character vector) Name of cells in the group.}
}
\value{
Returns an URD object, with a column in \code{@group.ids}
marked \code{TRUE} for cells in \code{cells} and \code{FALSE} for
all other cells.
}
\description{
This function will create a column in \code{group.ids} that identifies
whether cells are part of a provided list of cells. This can be used to
highlight a group of cells in any of URD's plotting functions.
}
