% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-map.R
\name{importDM}
\alias{importDM}
\title{Import a pre-calculated diffusion map}
\usage{
importDM(object, dm, cell.names = NULL)
}
\arguments{
\item{object}{An URD object}

\item{dm}{A DiffusionMap object produced by destiny's \code{\link[destiny]{DiffusionMap}}.}
}
\value{
An URD object
}
\description{
Import a pre-calculated diffusion map
}
