% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.dderiv.double}
\alias{impulse.dderiv.double}
\title{Second derivative of double sigmoid ("impulse")}
\usage{
impulse.dderiv.double(x, b1, b2, h0, h1, h2, t1, t2)
}
\arguments{
\item{x}{(Numeric)}

\item{b1}{(Numeric) Rate parameter of first sigmoid (should be negative for impulse)}

\item{b2}{(Numeric) Rate parameter of second sigmoid (should be negative for impulse)}

\item{h0}{(Numeric) Initial level}

\item{h1}{(Numeric) Peak level}

\item{h2}{(Numeric) Final level}

\item{t1}{(Numeric) Midpoint of first sigmoid function}

\item{t2}{(Numeric) Midpoint of second sigmoid function}
}
\value{
Returns numeric vector of second derivative of double sigmoid ("impulse") with given parameters at each x.
}
\description{
Second derivative of double sigmoid ("impulse")
}
\keyword{internal}
