% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.dderiv.single}
\alias{impulse.dderiv.single}
\title{Second derivative of single sigmoid}
\usage{
impulse.dderiv.single(x, b1, h0, h1, t1)
}
\arguments{
\item{x}{(Numeric)}

\item{b1}{(Numeric) Rate parameter of sigmoid}

\item{h0}{(Numeric) Initial level}

\item{h1}{(Numeric) Final level}

\item{t1}{(Numeric) Midpoint of sigmoid function}
}
\value{
Returns numeric vector of second derivative of single sigmoid with given parameters at each x.
}
\description{
Second derivative of single sigmoid
}
\keyword{internal}
