% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.double}
\alias{impulse.double}
\title{Evaluate the impulse model (double sigmoid) function, given parameters}
\usage{
impulse.double(x, b1, b2, h0, h1, h2, t1, t2)
}
\arguments{
\item{x}{(Numeric)}

\item{b1}{(Numeric) slope of onset (should be negative for impulse)}

\item{b2}{(Numeric) slope of offset (should be positive for impulse)}

\item{h0}{(Numeric) initial level}

\item{h1}{(Numeric) peak level}

\item{h2}{(Numeric) after offset level ("steady-state")}

\item{t1}{(Numeric) time of onset}

\item{t2}{(Numeric) time of offset}
}
\value{
Double-sigmoid impulse function with parameters b1, b2, h0, h1, h2, t1, t2, as evaluated for the vector x.
}
\description{
Chechik et al. 2008 Nat Biotech, with Michal's help
}
