% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.fit.double}
\alias{impulse.fit.double}
\title{Fit Impulse Model (Double sigmoid)}
\usage{
impulse.fit.double(x, y, k = 20, limit.shape = c("none", "convex",
  "concave"))
}
\arguments{
\item{x}{(Numeric)}

\item{y}{(Numeric)}

\item{k}{(Numeric) Number of starting conditions to try}

\item{limit.shape}{("none", "convex", "concave")}
}
\description{
Fit Impulse Model (Double sigmoid)
}
