% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.llrtest}
\alias{impulse.llrtest}
\title{Log likelihood ratio test for impulse fits}
\usage{
impulse.llrtest(x, y, I1, I2, sd.bg, df = 3)
}
\arguments{
\item{x}{(Numeric) x-values}

\item{y}{(Numeric)}

\item{I1}{(Numeric)}

\item{I2}{(Numeric)}

\item{sd.bg}{(Numeric) Standard deviation of background data for estimating noise model}

\item{df}{(Numeric) Difference in degrees of freedom (i.e. number of parameters)}
}
\description{
Determines whether additional degrees of freedom in double sigmoid impulse fit are
warranted by improved fit to the data
}
