% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.single}
\alias{impulse.single}
\title{Evaluate the impulse model (single sigmoid) function, given parameters}
\usage{
impulse.single(x, b1, h0, h1, t1)
}
\arguments{
\item{x}{(Numeric)}

\item{b1}{(Numeric) slope of onset (should be negative for onset)}

\item{h0}{(Numeric) initial level}

\item{h1}{(Numeric) after onset level ("steady-state")}

\item{t1}{(Numeric) time of onset}
}
\value{
Double-sigmoid impulse function with parameters b1, b2, h0, h1, h2, t1, t2, as evaluated for the vector x.
}
\description{
Chechik et al. 2008 Nat Biotech, with Michal's help
}
