% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.start.double}
\alias{impulse.start.double}
\title{Starting Conditions for impulse model (double sigmoid)}
\usage{
impulse.start.double(x, y, k, limit.shape = c("none", "concave", "convex"))
}
\description{
Function to determine k sets of initial starting conditions for impulse fitting
}
