% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dim-3D.R
\name{interpolate.points}
\alias{interpolate.points}
\title{Interpolate dimensionality reduction dimensions}
\usage{
interpolate.points(
  raw.coordinates,
  coord.prefix = "DC",
  dim.1,
  dim.2,
  dim.3,
  w.1 = rep(1/length(dim.1), length(dim.1)),
  w.2 = rep(1/length(dim.2), length(dim.2)),
  w.3 = rep(1/length(dim.3), length(dim.3)),
  rows.subset = NULL
)
}
\arguments{
\item{raw.coordinates}{(data.frame) Rows are cells, and columns are dimensions to use for interpolation.}

\item{coord.prefix}{(Character) Append this prefix to the dimensions}

\item{dim.1}{(Character vector) Dimensions to interpolate as dimension 1}

\item{dim.2}{(Character vector) Dimensions to interpolate as dimension 2}

\item{dim.3}{(Character vector) Dimensions to interpolate as dimension 3}

\item{w.1}{(Numeric vector) Weight to use for the members of \code{dim.1}. Defaults to equal weights for all members. Important to note that these can be negative to reverse the contribution of particular dimensions.}

\item{w.2}{(Numeric vector) Weight to use for the members of \code{dim.2}. Defaults to equal weights for all members. Important to note that these can be negative to reverse the contribution of particular dimensions.}

\item{w.3}{(Numeric vector) Weight to use for the members of \code{dim.3}. Defaults to equal weights for all members. Important to note that these can be negative to reverse the contribution of particular dimensions.}

\item{rows.subset}{(Character vector) Which rows to subset before interpolating. (Default \code{NULL} uses all rows.)}
}
\value{
nrows x 3 data.frame, where > 3 dimensions have been interpolated to produce a 3-dimensional embedding.
}
\description{
Allows for fitting more than three dimensions into a dimensionality reduction plot.
}
\keyword{internal}
