% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{inv.logistic}
\alias{inv.logistic}
\title{Inverse logistic function}
\usage{
inv.logistic(x, x0, k, c = 1)
}
\arguments{
\item{x}{(Numeric vector) Input values}

\item{x0}{(Numeric) Inflection point}

\item{k}{(Numeric) Slope}

\item{c}{(Numeric) Max value (default: 1)}
}
\description{
Inverse logistic function
}
\keyword{internal}
