% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check whether input is a whole number}
\usage{
is.wholenumber(x)
}
\arguments{
\item{x}{(Numeric vector) Input value(s)}
}
\description{
Check whether input is a whole number
}
