% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{llrtest.dof}
\alias{llrtest.dof}
\title{Log likelihood ratio test for impulse fits}
\usage{
llrtest.dof(y, M0, M, sd.bg, df)
}
\arguments{
\item{y}{(Numeric) Values of fit data}

\item{M0}{(Numeric) Values from fit function M0 (null)}

\item{M}{(Numeric) Values from fit function M (test)}

\item{sd.bg}{(Numeric) Standard deviation of background data for estimating noise model}

\item{df}{(Numeric) Difference in degrees of freedom (i.e. number of parameters) between M and M0.}
}
\description{
Determines whether additional degrees of freedom in double sigmoid impulse fit are
warranted by improved fit to the data
}
