% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-tips.R
\name{loadTipCells}
\alias{loadTipCells}
\title{Load Tip Cells}
\usage{
loadTipCells(object, tips)
}
\arguments{
\item{object}{An URD object}

\item{tips}{(Character) A clustering that was used as the tips (i.e. a column name of \code{@group.ids})}
}
\value{
An URD object with the cells that were used to start each tip stored in \code{@tree$cells.in.tip}
}
\description{
Load Tip Cells
}
