% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{log.likelihood.ratio}
\alias{log.likelihood.ratio}
\title{Log-likelihood ratio test}
\usage{
\method{log}{likelihood.ratio}(D, M0, M, sd.bg)
}
\arguments{
\item{D}{(Numeric)}

\item{M0}{(Numeric)}

\item{M}{(Numeric)}

\item{sd.bg}{(Numeric) Standard deviation of background genes for calculating null}
}
\description{
Log-likelihood ratio test
}
