% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-exp.R
\name{markersAUCPR}
\alias{markersAUCPR}
\title{Test for markers of a population using a precision-recall curve}
\usage{
markersAUCPR(
  object,
  clust.1 = NULL,
  clust.2 = NULL,
  cells.1 = NULL,
  cells.2 = NULL,
  clustering = NULL,
  effect.size = 0.25,
  frac.must.express = 0.1,
  exp.thresh = 0,
  frac.min.diff = 0,
  auc.factor = 1,
  max.auc.threshold = 1,
  genes.use = NULL
)
}
\arguments{
\item{object}{An URD object}

\item{clust.1}{(Character)}

\item{clust.2}{(Character)}

\item{cells.1}{(Character vector)}

\item{cells.2}{(Character vector)}

\item{clustering}{(Character) Name of a clustering (i.e. a column in \code{@group.ids})}

\item{effect.size}{(Numeric) Minimum log fold-change for two genes to be considered differentially expressed}

\item{frac.must.express}{(Numeric) Gene must be expressed in at least this fraction of cells in one of the two clusters to be considered.}

\item{exp.thresh}{(Numeric) Minimum expression value to consider 'expressed.'}

\item{frac.min.diff}{(Numeric) Fraction of cells expressing the gene must be at least this different between two populations to be considered.
@param auc.factor (Numeric) The precision-recall AUC is determined for a random classifier is determined based on the size of populations. To be considered differential, genes must have an AUC that is \code{auc.factor} times the expected AUC for a random classifier.}

\item{max.auc.threshold}{(Numeric) This acts as an upper bound for how high the AUC must be for a gene to be considered differential (i.e. you can choose that no matter the expectation for a random classifier, any gene with an AUC > 0.9 is considered differential).}

\item{genes.use}{(Character vector) Genes to compare, default is NULL (all genes)}
}
\value{
(data.frame)
}
\description{
Test for markers of a population using a precision-recall curve
}
