% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-smooth.R
\name{matrixReduce}
\alias{matrixReduce}
\title{Reduce matrix}
\usage{
matrixReduce(m, agg.func = mean)
}
\arguments{
\item{m}{(data.frame or Matrix) Input}

\item{agg.func}{(function) Function to use to aggregate columns with same
colname. Default is \code{mean}.}
}
\value{
(data.frame) Input with columns with same colname aggregated.
}
\description{
This takes a matrix/data.frame and combines all columns with the same column name
(for instance, the same pseudotime) using \code{agg.fun}, which defaults
to \code{mean}.
}
\keyword{internal}
