% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mean.of.logs.pos}
\alias{mean.of.logs.pos}
\title{Arithmetic mean of positive log-transformed values}
\usage{
mean.of.logs.pos(x, base = 2)
}
\arguments{
\item{x}{(Numeric vector) Input values}

\item{base}{(Numeric) Log base to use (default: 2)}
}
\description{
Un-logs, takes the arithmetic mean of those values greater than 0, and re-logs
}
