% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mean.of.logs}
\alias{mean.of.logs}
\title{Arithmetic mean of log-transformed values}
\usage{
mean.of.logs(x, base = 2)
}
\arguments{
\item{x}{(Numeric vector) Input values}

\item{base}{(Numeric) Log base to use (default: 2)}
}
\description{
Un-logs, takes the arithmetic mean, and re-logs
}
\keyword{internal}
