% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFDoublets.R
\name{NMFDoubletsDetermineCells}
\alias{NMFDoubletsDetermineCells}
\title{NMF Doublets: Identify cells that express non-overlapping NMF modules}
\usage{
NMFDoubletsDetermineCells(object, module.combos, module.expressed.thresh,
  frac.overlap.max, frac.overlap.diff.max)
}
\arguments{
\item{object}{An URD object}

\item{module.combos}{(List) Output from \code{\link{NMFDoubletsDefineModules}}}

\item{module.expressed.thresh}{(Numeric) Treshold for considering a module 'expressed' in a cell}

\item{frac.overlap.max}{(Numeric)}

\item{frac.overlap.diff.max}{(Numeric)}
}
\value{
(Character Vector) Cell Names that are potentially doublets
}
\description{
This uses NMF modules (as loaded into slot \code{@nmf.c1}) to determine cells
that express pairs of non-overlapping NMF modules and are likely to be technical
doublets.
}
\details{
#' This considers NMF modules (as loaded into slot \code{@nmf.c1}) pairwise
to determine how many cells have overlapping expression of that pair of
modules. Two thresholds are used to determine whether a cell 'expresses'
a module -- this helps differentiate between totally distinct modules (in
which case the number of co-expressing cells is similar with either threshold)
from pairs of modules expressed in an overlapping gradient (in which case
the number of cells that co-express them increases dramatically as the threshold
is lowered).
}
