% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFDoublets.R
\name{NMFDoubletsPlotModulesInCell}
\alias{NMFDoubletsPlotModulesInCell}
\title{Plot cells' expression of modules}
\usage{
NMFDoubletsPlotModulesInCell(object, modules, cells,
  module.expressed.thresh = Inf)
}
\arguments{
\item{object}{An URD object}

\item{modules}{(Character vector) Modules to plot (Currently column names of @gene.sig.z)}

\item{cells}{(Character vector) Cells to plot on y-axis}

\item{module.expressed.thresh}{(Numeric) Threshold for module expression. Will change the shape of points to emphasize whether they made cut-offs}
}
\value{
A ggplot2 object
}
\description{
For inspecting why particular cells do or don't get called as doublets
Looks for module expression in object@gene.sig.z for now
}
