% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaMarchenkoPastur}
\alias{pcaMarchenkoPastur}
\title{Marchenko-Pastur Significant PCs}
\usage{
pcaMarchenkoPastur(M, N, pca.sdev, do.print = T)
}
\arguments{
\item{M}{(Numeric) Number of rows in input data}

\item{N}{(Numeric) Number of columns in input data}

\item{pca.sdev}{(Numeric vector) Standard deviations for each principal component}

\item{do.print}{(Logical) Whether to report the results}
}
\value{
Logical vector of whether each PC is significant.
}
\description{
The Marchenko Pastur Law (MP) predicts the theoretical upper and lower bounds
on the null distribution of eigenvalues for an MxN random matrix. We take
significant principal components (PCs) as those with eigenvalues greater than
the maximum eigenvalue predicted for random data. This function assumes that
the data has mean 0 and variance 1 (i.e. that the data has been centered and
scaled). This called automatically by \code{\link{calc.PCA}} and the results
are stored in slot \code{pca.sig}.
}
