% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcSDPlot}
\alias{pcSDPlot}
\title{PC Standard Deviation Plot}
\usage{
pcSDPlot(object)
}
\arguments{
\item{object}{An URD object}
}
\value{
Nothing - produces a plot using R standard graphics.
}
\description{
Plots the standard deviation of each PC and the determined significant PC cut-off.
}
