% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcTopGenes}
\alias{pcTopGenes}
\title{PC Top Loaded Genes}
\usage{
pcTopGenes(object, pcs.print = which(object@pca.sig), genes.print = 20)
}
\arguments{
\item{object}{An URD object}

\item{pcs.print}{(Numeric vector) Which PCs to report? (Default: significant PCs)}

\item{genes.print}{(Numeric) How many genes to report}
}
\description{
PC Top Loaded Genes
}
