% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchpointPlot.R
\name{plotBranchpointByGene}
\alias{plotBranchpointByGene}
\title{Plot Gene Expression On Branchpoint Layout}
\usage{
plotBranchpointByGene(object, lineage.df, label, label.2 = NULL,
  label.type = "search", populations = NULL, visited.size = T,
  point.alpha = 0.2, pt.lim = NULL, color.scale = NULL,
  discrete.colors = NULL, ylab = "Pseudotime", xlab = "Preference",
  title = label, axis.lines = T, legend = T, fade.low = 0.5)
}
\arguments{
\item{object}{An URD object}

\item{lineage.df}{The output of \code{\link{branchpointPreferenceLayout}}}

\item{label}{(Character) Value to plot}

\item{label.2}{(Character) If dual-color plotting, the second value to plot.}

\item{label.type}{(Character)}

\item{populations}{(Character vector, length 2) Labels for the two populations on the x-axis.}

\item{visited.size}{(Logical) Should the size of points reflect how frequently they were visited by the random walks (to emphasize the more confident data points)}

\item{point.alpha}{(Numeric) Transparency of points}

\item{pt.lim}{(Numeric vector) Pseudotime (y) axis limits. Default \code{NULL} lets ggplot determine them automatically.}

\item{color.scale}{(Character vector) Vector of colors to use as a color scale if \code{label} is continuous values. Does not apply if a second label is given, in which case the color scale will be red-green.}

\item{discrete.colors}{(Character vector) Vector of colors to use as a color scale if \code{label} is discrete values.}

\item{ylab}{(Character) Y-axis label}

\item{xlab}{(Character) X-axis label}

\item{title}{(Character) Plot title}

\item{axis.lines}{(Logical) Plot axis lines on the plot?}

\item{legend}{(Logical) Plot an expression legend?}

\item{fade.low}{(Numeric) The transparency of the lowest expression points are 
decreased by this factor in order to prevent them from blocking the visualization 
of expressing cells if they are intermixed.}
}
\value{
A ggplot2 object
}
\description{
Plot Gene Expression On Branchpoint Layout
}
