% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dim-3D.R
\name{plotDim3DStoreView}
\alias{plotDim3DStoreView}
\title{Store a 3D view for plotDim3D}
\usage{
plotDim3DStoreView(
  object,
  view.name,
  reduction.use = c("dm", "pca"),
  dim.1 = 1,
  dim.2 = 2,
  dim.3 = 3,
  w.1 = NULL,
  w.2 = NULL,
  w.3 = NULL
)
}
\arguments{
\item{object}{An URD object}

\item{view.name}{(Character) What the name of the view should be.}

\item{reduction.use}{(Character) Dimensionality reduction to use (Diffusion Map ("dm") or PCA ("pca"))}

\item{dim.1}{(Numeric vector) Component(s) to use on x-axis. If > 1, they will be interpolated according to weights in \code{w.1}.}

\item{dim.2}{(Numeric vector) Component(s) to use on y-axis. If > 1, they will be interpolated according to weights in \code{w.2}.}

\item{dim.3}{(Numeric vector) Component(s) to use on z-axis. If > 1, they will be interpolated according to weights in \code{w.3}.}

\item{w.1}{(Numeric vector) Weight to use for the members of \code{dim.1}. Defaults to equal weights for all members. Important to note that these can be negative to reverse the contribution of particular dimensions.}

\item{w.2}{(Numeric vector) Weight to use for the members of \code{dim.1}. Defaults to equal weights for all members. Important to note that these can be negative to reverse the contribution of particular dimensions.}

\item{w.3}{(Numeric vector) Weight to use for the members of \code{dim.1}. Defaults to equal weights for all members. Important to note that these can be negative to reverse the contribution of particular dimensions.}
}
\value{
An URD object with a new 3D view stored in \code{@plot.3d}.
}
\description{
This allows storing "views" which contain the information about which dimensionality reduction,
dimensions, weights, and 3D orientation to use to make plots. This is useful for
generating figures with many similar panels.
}
