% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dim.R
\name{plotDimArray}
\alias{plotDimArray}
\title{Dimensionality Reduction Plot Array}
\usage{
plotDimArray(
  object,
  reduction.use = c("dm", "pca"),
  dims.to.plot,
  outer.title = NULL,
  file = NULL,
  file.width = 750,
  file.height = 600,
  ...
)
}
\arguments{
\item{object}{An URD object}

\item{reduction.use}{(Character) Dimensionality reduction to use}

\item{dims.to.plot}{(Numeric vector) Dimensions to plot. Will be plotted as pairs, with odd indices on the x-axes and even indices on the y-axes.}

\item{outer.title}{(Charater) A title to place outside the entire array of plots.}

\item{file}{(Character) Path to plot to save (if NULL, plot is returned)}
}
\value{
If \code{file=NULL}, returns a grid.array, otherwise returns nothing.
}
\description{
For surveying several dimensions of a dimensionality reduction (as in
\code{\link{plotDim}}). This will produce the same plot, but varying
the pair of dimensions shown. It optionally saves directly to a file,
since this can produce large plots.
}
