% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dim.R
\name{plotDimDual}
\alias{plotDimDual}
\title{Dimensionality Reduction Plot (Dual Color)}
\usage{
plotDimDual(object, label.red, label.green, label.red.type = "search",
  label.green.type = "search", reduction.use = c("tsne", "pca", "dm"),
  dim.x = 1, dim.y = 2, point.size = 1, alpha = 1, plot.title = "",
  legend = T, legend.size = 1/5.5, legend.offset.x = 0,
  legend.offset.y = 0, label.clusters = F, x.lim = NULL, y.lim = NULL)
}
\arguments{
\item{object}{An URD object}

\item{label.red}{(Character) Data to use for coloring points for the red channel}

\item{label.green}{(Character) Data to use for coloring points for the green channel}

\item{label.red.type}{(Character) Type of data to search for the label for the red channel. Default is "search" which checks several data types in order. For more information: \code{\link{data.for.plot}}}

\item{label.green.type}{(Character) Type of data to search for the label for the green channel. Default is "search" which checks several data types in order. For more information: \code{\link{data.for.plot}}}

\item{reduction.use}{(Character) Dimensionality reduction to use (tSNE, PCA, or Diffusion Map)}

\item{dim.x}{(Numeric) Component to use on x-axis}

\item{dim.y}{(Numeric) Component to use on y-axis}

\item{point.size}{(Numeric) Size of points on plot}

\item{alpha}{(Numeric) Transparency of points on plot: 0 (Transparent) - 1 (Opaque)}

\item{plot.title}{(Character) Title of the plot}

\item{legend}{(Logical) Show a legend?}

\item{legend.size}{(Numeric) Adjusts the size of the legend.}

\item{legend.offset.x}{(Numeric) Adjust the legend position (in terms of dimensionality reduction coordinates)}

\item{legend.offset.y}{(Numeric) Adjust the legend position (in terms of dimensionality reduction coordinates)}

\item{label.clusters}{(Logical) Label centroids of a discrete label?}

\item{x.lim}{(Numeric) Limits of x-axis (NULL autodetects)}

\item{y.lim}{(Numeric) Limits of y-axis (NULL autodetects)}

\item{colors}{(Character vector) Colors to use to generate a gradient scale to color continuous data}

\item{discrete.colors}{(Character vector) Colors to use to color}
}
\value{
A ggplot2 object
}
\description{
Plots cells according to their coordinates in a dimensionality reduction (tSNE by default).
Cells are colored according to two user set labels (\code{label.red} and \code{label.green})
in a mode that simulates dual-color microscopy - cells that express neither label are black,
cells that express one label are red or green, and cells that express both labels are yellow.
Both labels must be continuous variables (i.e. not cluster identities).
}
