% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dim.R
\name{plotDimHighlight}
\alias{plotDimHighlight}
\title{Dimensionality Reduction Plot With Highlighted Clusters}
\usage{
plotDimHighlight(object, clustering, cluster, highlight.color = "red", ...)
}
\arguments{
\item{object}{An URD object}

\item{clustering}{(Character) Name of column in \code{@group.ids} that identifies the clustering to pull from}

\item{cluster}{(Character vector) A cluster name that you want to highlight.}

\item{highlight.color}{(Character) A color name to use for the highlighted cluster.}

\item{...}{Additional parameters to pass to \code{\link{plotDim}}}
}
\value{
A ggplot2 object
}
\description{
Produces a plot with \code{\link{plotDim}} with cluster colors dimmed, and
a single cluster highlighted in a bright color (by default, red).
}
