% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-genes.R
\name{plotDists}
\alias{plotDists}
\title{Plot distributions}
\usage{
plotDists(object, label, category.label, label.type = "search",
  category.label.type = "search", legend = T, plot.title = "")
}
\arguments{
\item{object}{An URD object}

\item{label}{(Character) Data to use for distributions (e.g. "pseudotime", a metadata name, or a gene name). Must be continuous.}

\item{category.label}{(Character) Data to use for dividing into separate curves (e.g. developmental stage or a clustering). Must be discrete.}

\item{label.type}{(Character) Type of data to search for the label. Default is "search" which checks several data types in order. For more information: \code{\link{data.for.plot}}}

\item{category.label.type}{(Character) Type of data to search for category.label. Default is "search" which checks several data types in order. For more information: \code{\link{data.for.plot}}}

\item{legend}{(Logical) Show a legend?}

\item{plot.title}{(Character) Title of the plot}
}
\value{
A ggplot2 object. If used in a loop, plot(plotDists(...)) must be used.
}
\description{
This will plot the distribution of a continuous label across discrete classes.
For instance, it can be used to plot the distribution of pseudotime in each
developmental stage.
}
\examples{

}
