% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-genes.R
\name{plotScatter}
\alias{plotScatter}
\title{Gene Expression Scatterplot}
\usage{
plotScatter(object, label.x, label.y, label.color = NULL,
  label.x.type = "search", label.y.type = "search",
  label.color.type = "search", cells = NULL, point.size = 1,
  point.alpha = 1, xlim = NULL, ylim = NULL, density.background = T,
  density.color = "#888888", xlab = label.x, ylab = label.y,
  colorlab = label.color, title = "", legend = T)
}
\arguments{
\item{object}{An URD object}

\item{label.x}{(Character) Value to plot on the x-axis}

\item{label.y}{(Character) Value to plot on the y-axis}

\item{label.color}{(Character) Value to plot as point color}

\item{label.x.type}{(Character)}

\item{label.y.type}{(Character)}

\item{label.color.type}{(Character)}

\item{cells}{(Character vector) Names of cells to include in the plot (Default, \code{NULL} is all cells.)}

\item{point.size}{(Numeric) Adjust size of points on the plot}

\item{point.alpha}{(Numeric) Adjust transparency of points on the plot}

\item{xlim}{(Numeric vector) X-axis limits (Default \code{NULL} lets ggplot2 automatically decide.)}

\item{ylim}{(Numeric vector) Y-axis limits (Default \code{NULL} lets ggplot2 automatically decide.)}

\item{density.background}{(Logical) Should}

\item{density.color}{(Character) Darkest color of the density scale}

\item{xlab}{(Character) X-axis label}

\item{ylab}{(Character) Y-axis label}

\item{colorlab}{(Character) Color guide label}

\item{title}{(Character) Title of plot}

\item{legend}{(Logical) Display a legend}
}
\description{
Gene Expression Scatterplot
}
