% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-plots.R
\name{plotTree}
\alias{plotTree}
\title{Plot 2D Dendrogram of URD Tree}
\usage{
plotTree(object, label = NULL, label.type = "search", title = label,
  legend.title = "", plot.tree = T, tree.alpha = 1, tree.size = 1,
  plot.cells = T, cell.alpha = 0.25, cell.size = 0.5, label.x = T,
  label.segments = F, segment.colors = NULL, discrete.ignore.na = F,
  color.tree = T, hide.y.ticks = T)
}
\description{
Plot 2D Dendrogram of URD Tree
}
