% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-plots.R
\name{plotTreeDual}
\alias{plotTreeDual}
\title{Plot 2D Dendrogram of URD Tree}
\usage{
plotTreeDual(
  object,
  label.red,
  label.green,
  label.type.red = "search",
  label.type.green = "search",
  title = NULL,
  legend = T,
  legend.title = "",
  plot.tree = T,
  tree.alpha = 1,
  tree.size = 1,
  plot.cells = T,
  cell.alpha = 0.25,
  cell.size = 0.5,
  label.x = T,
  label.segments = F,
  color.tree = T,
  color.limits.red = NULL,
  color.limits.green = NULL,
  hide.y.ticks = T
)
}
\arguments{
\item{object}{An URD object}

\item{label.red}{(Character) Data to use for coloring points/tree in the red channel: see \link{data.for.plot}}

\item{label.green}{(Character) Data to use for coloring points/tree in the green channel: see \link{data.for.plot}}

\item{label.type.red}{(Character) Type of data to search for the label for the red channel. Default is "search" which checks several data types in order. For more information: \code{\link{data.for.plot}}}

\item{label.type.green}{(Character) Type of data to search for the label for the green channel. Default is "search" which checks several data types in order. For more information: \code{\link{data.for.plot}}}

\item{title}{(Character) Title to display on the plot. Default \code{NULL} auto-creates a title based on what is plotted.}

\item{legend}{(Logical) Show a legend?}

\item{legend.title}{(Character) Title to display on the legend}

\item{plot.tree}{(Logical) Whether cells should be plotted with the tree}

\item{tree.alpha}{(Numeric) Transparency of dendrogram (0 is transparent, 1 is opaque)}

\item{tree.size}{(Numeric) Thickness of lines of dendrogram}

\item{cell.alpha}{(Numeric) Transparency of cells (0 is transparent, 1 is opaque)}

\item{cell.size}{(Numeric) How large should cells be}

\item{label.x}{(Logical) Should tips on the x-axis be labeled}

\item{label.segments}{(Logical) Should segments of the dendrogram be labeled with their numbers}

\item{color.tree}{(Logical) Should the dendrogram be colored according to the data?}

\item{color.limits.red}{(Numeric vector, length 2) Minimum and maximum values for color scale. Default \code{NULL} auto-detects.}

\item{color.limits.green}{(Numeric vector, length 2) Minimum and maximum values for color scale. Default \code{NULL} auto-detects.}

\item{hide.y.ticks}{(Logical) Should the pseudotime values on the y-axis be hidden?}
}
\value{
A ggplot2 object
}
\description{
Plot 2D Dendrogram of URD Tree
}
