% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{plotTreeForce2D}
\alias{plotTreeForce2D}
\title{Plot tree force-directed layout in 2D}
\usage{
plotTreeForce2D(object, label = NULL, label.type = "search",
  title = label, show.points = T, point.alpha = 1, point.size = 1,
  show.neighbors = F, neighbors.max = 10000, colors = NULL)
}
\arguments{
\item{object}{An URD object}
}
\description{
Plot tree force-directed layout in 2D
}
