% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{plotTreeForceStore3DView}
\alias{plotTreeForceStore3DView}
\title{Store a 3D view for force-directed layouts}
\usage{
plotTreeForceStore3DView(object, view.name)
}
\arguments{
\item{object}{An URD object}

\item{view.name}{(Character) Name to use for this view}
}
\value{
An URD object with the parameters for that view stored in \code{@tree$force.view.list}.
}
\description{
The last stored view will be the "default."
}
