% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-plots.R
\name{plotTreeHighlight}
\alias{plotTreeHighlight}
\title{Plot 2D Dendrogram of URD Tree (with cells that meet arbitrary criteria highlighted)}
\usage{
plotTreeHighlight(object, label.name, label.value, color = "red",
  bg.color = "#CCCCCC", highlight.alpha = 0.6, highlight.size = 1.5,
  combine = c("intersect", "union"), ...)
}
\arguments{
\item{object}{An URD object}

\item{label.name}{(Character vector) Data to use for selecting cells, see \link{data.for.plot}}

\item{label.value}{(List of vectors) List of same length as \code{label.name}; Each entry is the acceptable values for the corresponding entry in \code{label.name}}

\item{color}{(Character) Color to use for highlighted cells}

\item{bg.color}{(Character) Color to use for non-highlighted cells}

\item{highlight.alpha}{(Numeric) Transparency of highlighted cells (0 is fully transparent, 1 is fully opaque)}

\item{highlight.size}{(Numeric) Size of points of highlighted cells}

\item{combine}{(Character) Should highlighted cells be the \code{intersect} (i.e. cells that meet ALL criteria) or the \code{union} (i.e. cells that meet any criteria)}

\item{...}{all additional parameters are passed to \code{\link{plotTree}}}
}
\value{
A ggplot2 object
}
\description{
This produces an URD dendrogram with cells that meet any group of arbitrary criteria highlighted on the tree. It uses the \code{highlight.cells}, \code{highlight.cells.alpha}, and \code{highlight.cells.size} parameters of \code{\link{plotTree}} which can alternatively be used to accomplish this.
}
