% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-genes.R
\name{plotViolin}
\alias{plotViolin}
\title{Violin plot of gene expression}
\usage{
plotViolin(
  object,
  labels.plot,
  clustering,
  clusters.use = NULL,
  legend = T,
  free.axes = F,
  point.size = 0.2,
  point.color = "black",
  point.alpha = 0.5
)
}
\arguments{
\item{object}{An URD object}

\item{labels.plot}{(Character vector) Data to use for coloring points (e.g. a metadata name, group ID from clustering, or a gene name)}

\item{clustering}{(Character) Name of clustering to use (i.e. a column name of \code{@group.ids})}

\item{clusters.use}{(Character vector) Names of clusters to display (default NULL uses all clusters)}

\item{legend}{(Logical) Display a legend?}

\item{free.axes}{(Logical) Should the scale of each label be individually determined?}
}
\value{
A ggplot2 object
}
\description{
Violin plot of gene expression
}
