% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pmax.abs}
\alias{pmax.abs}
\title{Return the signed value with greater magnitude}
\usage{
pmax.abs(x, y)
}
\arguments{
\item{x}{(Numeric vector) Input values}

\item{y}{(Numeric vector) Input values}
}
\description{
Determines which value has greater magnitude (i.e. by absolute value) and returns the original signed value.
Compares vectors x and y pairwise.
}
