% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{preference}
\alias{preference}
\title{Determining preference between a pair of values}
\usage{
preference(x, y, signed = F)
}
\arguments{
\item{x}{(Numeric vector)}

\item{y}{(Numeric vector)}

\item{signed}{(Logical)}
}
\description{
Determining preference between a pair of values
}
