% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{processRandomWalksFromTips}
\alias{processRandomWalksFromTips}
\title{Process random walks from all tips}
\usage{
processRandomWalksFromTips(object, walks.list, aggregate.fun = mean,
  n.subsample = 10, verbose = T)
}
\arguments{
\item{object}{An URD object}

\item{walks.list}{(List of lists) List of character vectors of cells visited during random walks.}

\item{aggregate.fun}{(Function) Function to aggregate pseudotime (default: mean)}

\item{n.subsample}{(Numeric) Number of subsamplings to perform for calculating}

\item{verbose}{(Logical) Report on progress}
}
\value{
An URD object with cell visitation frequency stored in \code{@diff.data}, a calculated pseudotime stored in \code{@pseudotime}, and
subsampled data in \code{@pseudotime.stability}.
}
\description{
Process random walks from all tips
}
