% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{prop.exp}
\alias{prop.exp}
\title{Returns proportion of expressing cells}
\usage{
prop.exp(x, exp.thresh = 0)
}
\arguments{
\item{x}{(Numeric vector) Input values}

\item{exp.thresh}{(Numeric) Minimum value considered expressed}
}
\description{
Determines proportion of input values > exp.thresh.
}
\keyword{internal}
