% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{prop.nonexp}
\alias{prop.nonexp}
\title{Returns proportion of non-expressing cells}
\usage{
prop.nonexp(x)
}
\arguments{
\item{x}{(Numeric vector) Input values}
}
\description{
Determines proportion of input values <= 0.
}
