% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-structure.R
\name{pseudotimeBreakpointByStretch}
\alias{pseudotimeBreakpointByStretch}
\title{Find Pseudotime Breakpoint}
\usage{
pseudotimeBreakpointByStretch(div.pseudotime, segment.1, segment.2, visit.data,
  pseudotime.windows, verbose = T)
}
\description{
Find Pseudotime Breakpoint
}
