% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-structure.R
\name{pseudotimeBreakpointByStretchV1}
\alias{pseudotimeBreakpointByStretchV1}
\title{Find Pseudotime Breakpoint}
\usage{
pseudotimeBreakpointByStretchV1(div.pseudotime, segment.1, segment.2,
  visit.data, pseudotime.windows, verbose = T)
}
\description{
Find Pseudotime Breakpoint
}
