% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{pseudotimeDetermineLogistic}
\alias{pseudotimeDetermineLogistic}
\title{Determine logistic parameters}
\usage{
pseudotimeDetermineLogistic(object, pseudotime, optimal.cells.forward,
  max.cells.back, pseudotime.direction = ">", asymptote = 0.01,
  do.plot = T, print.values = T)
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character)}

\item{optimal.cells.forward}{(Numeric)}

\item{max.cells.back}{(Numeric)}

\item{pseudotime.direction}{(Character: "<" or ">")}

\item{asymptote}{(Numeric)}

\item{do.plot}{(Logical)}

\item{print.values}{(Logical)}
}
\value{
Logistic parameters
}
\description{
Determine logistic parameters
}
