% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-cascade.R
\name{pseudotimeMovingWindow}
\alias{pseudotimeMovingWindow}
\title{Moving window through pseudotime}
\usage{
pseudotimeMovingWindow(object, pseudotime, cells, moving.window,
  cells.per.window, name.by = c("mean", "min", "max"))
}
\description{
Moving window through pseudotime
}
