% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime-stability-plots.R
\name{pseudotimePlotStabilityCells}
\alias{pseudotimePlotStabilityCells}
\title{Plot Pseudotime Stability (Individual Cells)}
\usage{
pseudotimePlotStabilityCells(object, cells.plot = 9, log.visits = T)
}
\arguments{
\item{object}{An URD object}

\item{cells.plot}{(Character vector or Numeric) Either names of cells to plot, or a number of cells to plot (chosen randomly).}

\item{log.visits}{(Logical) Whether to log10 transform visitation frequency}
}
\value{
A ggplot2 object
}
\description{
Plot Pseudotime Stability (Individual Cells)
}
