% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime-stability-plots.R
\name{pseudotimePlotStabilityOverall}
\alias{pseudotimePlotStabilityOverall}
\title{Plot Pseudotime Stability}
\usage{
pseudotimePlotStabilityOverall(object)
}
\arguments{
\item{object}{An URD object}
}
\value{
A ggplot2 object
}
\description{
Plot Pseudotime Stability
}
