% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime-stability-plots.R
\name{pseudotimePlotVisits}
\alias{pseudotimePlotVisits}
\title{Plot Cell Visitation By Random Walks or Flood Pseudotime}
\usage{
pseudotimePlotVisits(object, xlim = NULL)
}
\arguments{
\item{object}{An URD object}

\item{xlim}{(Numeric vector, length 2) Limits of histogram x-axis. (NULL allows ggplot to determine automatically.)}
}
\value{
A ggplot2 object
}
\description{
Plot Cell Visitation By Random Walks or Flood Pseudotime
}
