% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{pseudotimeWeightTransitionMatrix}
\alias{pseudotimeWeightTransitionMatrix}
\title{Weight transition matrix by pseudotime}
\usage{
pseudotimeWeightTransitionMatrix(object, pseudotime, x0 = NULL, k = NULL,
  logistic.params = NULL, pseudotime.direction = ">")
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character)}

\item{x0}{(Numeric)}

\item{k}{(Numeric)}

\item{logistic.params}{(List)}

\item{pseudotime.direction}{(Character: ">" or "<")}
}
\value{
Sparse Matrix (dgCMatrix) of transition probabilities, weighted by pseudotime
}
\description{
Weight transition matrix by pseudotime
}
