% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{rotateCoords3d}
\alias{rotateCoords3d}
\title{Rotate coordinates in 3D}
\usage{
rotateCoords3d(m, a, axis = c("x", "y", "z"))
}
\arguments{
\item{m}{(Matrix) n cells x 3 matrix of coordinates}

\item{a}{(Numeric) Angle to rotate}

\item{axis}{(Character) Which axis to rotate around}
}
\value{
n x 3 matrix of rotated coordinates
}
\description{
Rotate coordinates in 3D
}
\keyword{internal}
