% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-navigate.R
\name{segChildrenAll}
\alias{segChildrenAll}
\title{All children of segment}
\usage{
segChildrenAll(object, segment, include.self = F, original.joins = F,
  point.at.segment.joins.initial = original.joins)
}
\description{
All children of segment
}
