% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-navigate.R
\name{segTerminal}
\alias{segTerminal}
\title{Get all terminal segments from a tree}
\usage{
segTerminal(object)
}
\arguments{
\item{object}{An URD object}
}
\value{
A character vector of segment names
}
\description{
Identify the segments that do not have any descendants in URD's determined
tree structure
}
