% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{seuratToURD}
\alias{seuratToURD}
\title{Import URD from Seurat}
\usage{
seuratToURD(seurat.object)
}
\arguments{
\item{seurat.object}{A Seurat object}
}
\value{
An URD object
}
\description{
If you're previously performed analyses in Seurat, you can copy them over
directly into an URD object for analyzing pseudotime, pseudolineage, and
building a specification tree.
}
