% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urd-class.R
\docType{methods}
\name{show}
\alias{show}
\title{URD show method}
\usage{
\S4method{show}{URD}(object)
}
\arguments{
\item{object}{An URD object}
}
\description{
Prevents R from crashing by trying to print all slots of an URD object
if a returned object is not stored in a variable.
}
