% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparseApply}
\alias{sparseApply}
\title{Equivalent of apply function for sparse matrices from Matrix}
\usage{
sparseApply(x, margin, fun, max.dim, verbose = F, ...)
}
\arguments{
\item{x}{(dgCMatrix) Input}

\item{margin}{(Numeric) \code{1} for rows, \code{2} for columns}

\item{fun}{(Function)}

\item{max.dim}{(Numeric) Number of rows or columns to process at a time}

\item{verbose}{(Logical) Report progress reports on chunks}

\item{...}{Additional arguments to pass to \code{fun}}
}
\description{
Matrix package will 'automagically' and silently convert dgCMatrix sparse matrices to a
dense matrix when they are passed to the apply function. While rowMeans, colMeans, rowSums,
and colSums functions exist to reduce need for apply function, sometimes other functions
are needed. For loops are remarkably inefficient, so this will chunk the original matrix into
pieces, process them using apply, convert the results back to sparse if needed, then assemble
into a final result and return. It's a mediocre balance between memory usage and speed.
}
\keyword{internal}
