% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{sum.of.logs}
\alias{sum.of.logs}
\title{Sum of log-transformed values}
\usage{
\method{sum}{of.logs}(x, base = 2)
}
\arguments{
\item{x}{(Numeric vector) Input values}

\item{base}{(Numeric) Log base to use (default: 2)}
}
\description{
Un-logs, takes the sum, and re-logs
}
