% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{translateCoords3d}
\alias{translateCoords3d}
\title{Translate coordinates in 3D}
\usage{
translateCoords3d(m, x, y, z)
}
\arguments{
\item{m}{(Matrix) n cells x 3 coordinates matrix}

\item{x}{(Numeric) Distance to move in x}

\item{y}{(Numeric) Distance to move in y}

\item{z}{(Numeric) Distance to move in z}
}
\value{
n x 3 matrix of translated coordinates
}
\description{
Translate coordinates in 3D
}
\keyword{internal}
