% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-navigate.R
\name{translateSegmentNames}
\alias{translateSegmentNames}
\title{Translate Segment Names}
\usage{
translateSegmentNames(object, segments)
}
\arguments{
\item{object}{An URD object}

\item{segments}{(Character vector) A vector of segment names for which to return the segment numbers.}
}
\value{
Character vector of segment numbers.
}
\description{
Allows access to segments by name
}
