% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{treeForceRotateCoords}
\alias{treeForceRotateCoords}
\title{Rotate points in force-directed layout}
\usage{
treeForceRotateCoords(
  object,
  cells = NULL,
  seg = NULL,
  angle,
  axis = c("x", "y", "z"),
  around.cell = NULL,
  throw.out.cells = 0,
  pseudotime = NULL
)
}
\arguments{
\item{object}{An URD object}

\item{cells}{(Character vector) Cells to modify (Default \code{NULL} is all cells in the force-directed layout)}

\item{seg}{(Character) Instead of specifying cells, just grab all cells from this segment and downstream. Ignored if \code{cells} is specified.}

\item{angle}{(Numeric) Angle to rotate}

\item{axis}{(Character: "x", "y", "z") Axis around which to rotate}

\item{around.cell}{(Character or Numeric) If a character, then the name of a cell to rotate around. If numeric, then will choose the Nth cell by pseudotime. (This can be helpful for rotating a segment around its base to spread segments of the tree apart more. If \code{NULL}, then rotates around (0,0,0).}

\item{throw.out.cells}{(Numeric) Throw out the youngest N cells in pseudotime}

\item{pseudotime}{(Character) Pseudotime (i.e. column name of \code{@pseudotime}) to use for determining cells to throw out.}
}
\value{
An URD object with the coordinates of some cells in \code{@tree$walks.force.layout} modified.
}
\description{
This function can be used in addition to \code{\link{treeForceTranslateCoords}}
in order to fine-tune the presentation of a force-directed layout. For instance,
this can be used to achieve an improved 2D visualization by modifying overlapping
portions of the layout.
}
