% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{treeForceStretchCoords}
\alias{treeForceStretchCoords}
\title{Stretch points in force-directed layout}
\usage{
treeForceStretchCoords(
  object,
  factor,
  cells = NULL,
  seg = NULL,
  stretch.from = 20,
  throw.out.cells = 0,
  ramp = 0.1
)
}
\arguments{
\item{object}{An URD object}

\item{factor}{(Numeric) Factor to lengthen (>1) or shrink (<1) the branch by.}

\item{cells}{(Character vector) Cells to modify (Default \code{NULL} is all cells in the force-directed layout)}

\item{seg}{(Character) Instead of specifying cells, just grab all cells from this segment and downstream. Ignored if \code{cells} is specified.}

\item{throw.out.cells}{(Numeric) Throw out the youngest N cells in pseudotime (i.e. don't stretch those)}
}
\value{
An URD object with the coordinates of some cells in \code{@tree$walks.force.layout} modified.
}
\description{
This function can be used in addition to \code{\link{treeForceRotateCoords}}
and \code{\link{treeForceTranslateCoords}}
in order to fine-tune the presentation of a force-directed layout. For instance,
this can be used to achieve an improved 2D visualization by shortening or lengthening
a branch of the tree if branches have dramatically different lengths in pseudotime.
(This occurs, for instance, in data sets generated from a single timepoint in 
a homeostatic animal instead of an explicit timecourse.) This transformation
will only occur in the force-directed layout, not the dendrogram tree layout.
}
