% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-layout.R
\name{treeLayoutCells}
\alias{treeLayoutCells}
\title{Add cells to tree layout}
\usage{
treeLayoutCells(object, pseudotime, jitter = 0.15, jitter.push = 0.05)
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character) Pseudotime to use (i.e. a column name of \code{object@pseudotime}).}

\item{jitter}{(Numeric) Distance to jitter cells before placing next to dendrogram. (Relative to 1, which is the distance between terminal segments).}

\item{jitter.push}{(Numeric) Distance next to lines of dendrogram that should not include any cells.}
}
\value{
An URD object with cell locations added to \code{object@tree$cell.layout}.
}
\description{
This function is called automatically by \code{\link{buildTree}}.
}
\keyword{internal}
