% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-layout.R
\name{treeLayoutDendrogram}
\alias{treeLayoutDendrogram}
\title{Layout Tree Dendrogram}
\usage{
treeLayoutDendrogram(object)
}
\arguments{
\item{object}{An URD object}
}
\value{
An URD object with a dendrogram layout of segments stored in \code{object@tree$segment.layout}.
}
\description{
This function is called automatically by \code{\link{buildTree}}.
}
\keyword{internal}
