% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-layout.R
\name{treeLayoutElaborate}
\alias{treeLayoutElaborate}
\title{Elaborate Tree Dendrogram With Nodes}
\usage{
treeLayoutElaborate(object)
}
\arguments{
\item{object}{An URD object}
}
\value{
An URD object with a dendrogram that includes a short segment for
each node stored in \code{object@tree$tree.layout}.
}
\description{
This function is called automatically by \code{\link{buildTree}}.
}
