% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{txBarcodesIntegerToChar}
\alias{txBarcodesIntegerToChar}
\title{Convert 10X Integer-stored barcodes to character ones for subsetting matrices}
\usage{
txBarcodesIntegerToChar(x)
}
\arguments{
\item{x}{(Integer64 vector) 10X cell barcodes, stored as 64-bit integers, as is molecule_info.h5}
}
\value{
(Character vector) 10X cell barcodes, as DNA base characters
}
\description{
Are you kidding me with this? How much space does this save, guys?
}
