% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{txCutoffPlot}
\alias{txCutoffPlot}
\title{Plot Cumulative Distribution of Reads per Cell Barcode}
\usage{
txCutoffPlot(
  tx.reads.per.cell = NULL,
  molecule.info.path = NULL,
  n.cells = NULL,
  main = "",
  xmax = 5000
)
}
\arguments{
\item{n.cells}{(Numeric) Draw a line at putative number of cells to use downstream}

\item{main}{(Character) Title for the plot}

\item{xmax}{(Numeric) Max value of plot x-axis}

\item{gz.in}{(Character) Path to Dropseq cell read counts report (".report_cell_readcounts.txt.gz")}
}
\description{
To determine the number of cells recovered in a Dropseq sample, the cumulative
distribution of the number of reads recovered per unique cell barcode is plotted.
The elbow in the curve suggests the boundary between "real cells" with many reads
and "ambient RNA" that has bound to the remainder of the beads (which have few reads).
The 10X pipeline determines which barcodes constitute 'cells' in a different fashion,
but this produces an equivalent
}
