% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{txReadDGE}
\alias{txReadDGE}
\title{Read in Digital Gene Expression Matrix from 10X Cellranger}
\usage{
txReadDGE(output.path, dge.name = NULL, cells.read = NULL,
  name.by = c("gene", "id"), strip.num = T)
}
\arguments{
\item{output.path}{(Character) Path that contains barcodes.tsv, genes.tsv, and matrix.mtx from Cellranger}

\item{dge.name}{(Character) Name to append to all cell barcodes if multiple tables will be combined (default: NULL)}

\item{cells.read}{(Character vector) Cell names to trim the matrix to, if desired (will be matched after strip.num, before appending dge.name)}

\item{name.by}{(Character: "gene" or "id") Whether to name genes by gene ID, or gene name. If gene name, expression of gene IDs with the same name will be summed.}

\item{strip.num}{(Logical) Cellranger appends a batch number (e.g. "-1") to cell barcodes. If TRUE, this strips it away.}
}
\value{
A dgCMatrix, where rows are genes and columns are cells.
}
\description{
Read in Digital Gene Expression Matrix from 10X Cellranger
}
