% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropseq-dge.R
\name{txReadsPerCell}
\alias{txReadsPerCell}
\title{Read 10X Reads per Cell barcode}
\usage{
txReadsPerCell(molecule.info.path)
}
\arguments{
\item{molecule.info.path}{(Character) Path to the molecule_info.h5 file output by 10X.}
}
\value{
Returns a data.frame with columns "barcode" (2-bit encoded cell barcode), 
"reads", "unmapped_reads", and "nonconf_mapped_reads" (as per 10X documentation),
and "total_reads", which is the sum of the 3 previous columns.
}
\description{
This reads and aggregates the molecule info file output by the 10X pathway
into reads per cell barcode for use in \code{\link{txCutoffPlot}}. (This is
because it takes some time to read in this data, so many plots can be tried
with only having to read the data once.)
}
