% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{urdSubset}
\alias{urdSubset}
\title{Subset an URD object}
\usage{
urdSubset(object, cells.keep)
}
\arguments{
\item{object}{An URD object}

\item{cells.keep}{(Character vector) Names of cells to retain}
}
\value{
An URD object
}
\description{
This subsets an URD object, given a list of cells. This subsets all results
without recalculating them, which could produce non-sensical values for many
things. (For instance, it would make more sense to recalculate the PCA on
the subset data, rather than keeping previous PCA scores for only a subset
of cells.)
}
