% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cells.R
\name{whichCells}
\alias{whichCells}
\title{Which cells}
\usage{
whichCells(object, label, value, label.type = "search")
}
\arguments{
\item{object}{An URD object}

\item{label}{(Character) The label of the data to search for}

\item{value}{(Character or numeric vector) If \code{label} is discrete (e.g. a clustering, developmental stage, treatment), one or more character values to match against. If \code{label} is continuous (e.g. pseudotime), should be a numeric vector or length 2 that describes the inclusive range of acceptable values.}

\item{label.type}{(Character) Where to look for the data. Default is "search" which looks in order: "meta", "group", "sig", "gene", "counts", "pseudotime", "pca", "diff.data"}
}
\value{
(Character vector) Cell  that match the criteria specified.
}
\description{
Identifies cells that meet particular criteria and returns their cell names. This
can be used to select populations for subsetting, differential expression, plotting, etc.
}
\examples{
# Find cells from dome stage
whichCells(urd.object, "STAGE", "ZFDOME")

# Find cells with 500-1000 genes detected
whichCells(urd.object, "NUM_GENES", c(500,1000))

# Find cells with pseudotime between 0.54-0.72
whichCells(urd.object, "pseudotime", c(0.54,0.72))

}
