% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{buildTree}
\alias{buildTree}
\title{Build Tree}
\usage{
buildTree(object, pseudotime, tips.use = NULL, divergence.method = c("ks",
  "preference"), weighted.fusion = T, use.only.original.tips = T,
  cells.per.pseudotime.bin = 80, bins.per.pseudotime.window = 5,
  minimum.visits = 10, visit.threshold = 0.7,
  save.breakpoint.plots = NULL, save.all.breakpoint.info = F,
  p.thresh = 0.01, min.cells.per.segment = 1,
  min.pseudotime.per.segment = 0.01, dendro.node.size = 100,
  dendro.cell.jitter = 0.15, dendro.cell.dist.to.tree = 0.05, verbose = T)
}
\arguments{
\item{object}{An URD object}

\item{pseudotime}{(Character) Pseudotime to use for building tree}

\item{tips.use}{(Character vector) Name of tips to use in the final tree (Default is NULL, which uses all tips.)}

\item{divergence.method}{(Character: "ks" or "preference") Test to use to determine whether visitation has diverged for each pseudotime window.}

\item{cells.per.pseudotime.bin}{(Numeric) Approximate number of cells to assign to each pseudotime bin for branchpoint finding.}

\item{bins.per.pseudotime.window}{(Numeric) Width of moving window in pseudotime used for branchpoint finding, in terms of bins.}

\item{minimum.visits}{(Numeric) Minimum number of random walk visits to a cell to retain it in the tree}

\item{visit.threshold}{(Numeric) Cells are considered potential members for segments/tips from which random walks visited them 
at least this fraction of their maximum visitation from a single tip}

\item{save.breakpoint.plots}{(Path) Path to save plots summarizing (default is NULL, which does not save plots as they are somewhat slow)}

\item{save.all.breakpoint.info}{(Logical) Should all information about breakpoints be stored in the object for use in diagnostic plots? (Can add several hundred MB to object size.)}

\item{p.thresh}{(Numeric) p-value threshold to use in determining whether visitation is significantly different from pairs of tips}

\item{min.cells.per.segment}{(Numeric) Segments with fewer assigned cells weill be collapsed during tree construction}

\item{min.pseudotime.per.segment}{(Numeric) Segments shorter than this in pseudotime will be collapsed during tree construction}

\item{dendro.node.size}{(Numeric) Number of cells to assign per node (used for averaging expression for tree dendrogram branch coloring)}

\item{dendro.cell.jitter}{(Numeric) For the dendrogram tree layout, how much jitter to apply to cells. This can be revised after building the tree by re-generating the cell layout using \code{\link{treeLayoutCells}} with a different parameter.}

\item{dendro.cell.dist.to.tree}{(Numeric) For the dendrogram tree layout, how far to push cells away from the dendrogram branches. This can be revised after building the tree by re-generating the cell layout using \code{\link{treeLayoutCells}} with a different parameter.}

\item{verbose}{(Logical) Report on progress?}
}
\value{
An URD object with an URD-recovered tree structure stored in \code{@tree}
}
\description{
Build Tree
}
