% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-map.R
\name{calcDM}
\alias{calcDM}
\title{Calculate a diffusion map}
\usage{
calcDM(object, genes.use = object@var.genes, cells.use = NULL, knn = NULL,
  sigma.use = NULL, n_local = 5:7, distance = c("euclidean", "cosine",
  "rankcor"), density.norm = T, dcs.store = 200, verbose = T)
}
\arguments{
\item{object}{An URD object}

\item{genes.use}{(Character vector) Genes to use (default is variable genes, NULL uses all genes)}

\item{cells.use}{(Character vector) Cells to use (default is NULL, which uses all cells)}

\item{knn}{(Numeric) Number of nearest neighbors to use (NULL takes a guess)}

\item{sigma.use}{(NULL, numeric, or "local") Kernel width to use for the diffusion map (NULL uses destiny's global auto-detection procedure)}

\item{n_local}{(Numeric) If \code{sigma.use="local"}, the distance to the n-local(th) nearest neighbors are used for determining the kernel width for each cell.}

\item{distance}{(Character) Distance metric to use for determining transition probabilities.}

\item{density.norm}{(Logical) If TRUE, use density normalization}

\item{dcs.store}{(Numeric) How many diffusion components to store}

\item{verbose}{(Logical) Report determined values for auto-detected parameters?}
}
\description{
Calculate a diffusion map
}
