% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-smooth.R
\name{combineSmoothFit}
\alias{combineSmoothFit}
\title{Gene Smooth: Combine smoothed fits}
\usage{
combineSmoothFit(fit.list)
}
\arguments{
\item{fit.list}{(List) A list of smoothed.fits to combine}
}
\value{
(List) Akin to a single output from \code{\link{geneSmoothFit}}.
}
\description{
This function takes a list of smoothed.fits (i.e. output from \code{\link{geneSmoothFit}}) and combines them in the order provided into a single smoothed fit. This is useful, for instance, for combining multiple pieces of fits together for plotting on a heatmap with several panels.
}
