% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-map.R
\name{edgesFromDM}
\alias{edgesFromDM}
\title{Create an edge list from the transition matrix of the diffusion map}
\usage{
edgesFromDM(object, cells = NULL, edges.return = NULL)
}
\arguments{
\item{object}{An URD object}

\item{cells}{(Character vector) Cells to calculate edges between (default \code{NULL} uses all cells in diffusion map.)}

\item{edges.return}{(Numeric) Number of edges to return (default \code{NULL} returns all edges.)}
}
\value{
A data.frame with columns from (cell name), to (cell name), and weight (transition probability)
}
\description{
Returns a list of edges with transition probability from the diffusion
map in an URD object. Used by some plotting functions to show the
connectivity of the network if desired.
}
\examples{
edges <- edgesFromDM(object, cells=NULL, edges.return=100000)

}
