% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-cascade.R
\name{geneCascadeImpulsePlots}
\alias{geneCascadeImpulsePlots}
\title{Plot Impulse Fits For a Gene Cascade}
\usage{
geneCascadeImpulsePlots(cascade, genes = NULL, ymin0 = T, file = NULL)
}
\arguments{
\item{cascade}{(List) A gene cascade, such as generated by \code{\link{geneCascadeProcess}}}

\item{genes}{(Character vector) Genes to plot (default \code{NULL} plots all genes that were fit by \code{\link{geneCascadeProcess}})}

\item{ymin0}{(Logical) Should the minimum value of the y-axis be 0 (\code{TRUE}) or the minimum value of the fit data (\code{FALSE})}

\item{file}{(Path) A path to save the plot to as PDF (if NULL, display)}
}
\description{
This plots the results of fitting gene expression data with impulse models.
(i.e. from \code{\link{geneCascadeProcess}}.)
The windowed data is plotted as black points for each gene, and the fit curve
is plotted. Its color indicates the type of fit that
was chosen: blue (linear), green (single sigmoid), red (double sigmoid).
Diamonds indicate determined gene onset (orange) and gene offset (blue) times.
}
