% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-map.R
\name{importDM}
\alias{importDM}
\title{Import a pre-calculated diffusion map}
\usage{
importDM(object, dm, cell.names = NULL)
}
\arguments{
\item{object}{An URD object}

\item{dm}{A DiffusionMap object produced by destiny's \code{\link[destiny]{DiffusionMap}}.}
}
\value{
An URD object with the diffusion map stored in \code{@dm}.
}
\description{
If destiny is run separately (such as on a cluster to vary parameters), this
allows quick addition of destiny's output to an URD object.
}
\examples{
object <- importDM(object, dm.8)

}
