% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.fit.single}
\alias{impulse.fit.single}
\title{Fit Impulse Model (Single sigmoid)}
\usage{
impulse.fit.single(x, y, k = 20, limit.slope = c("none", "on", "off"))
}
\arguments{
\item{x}{(Numeric)}

\item{y}{(Numeric)}

\item{k}{(Numeric) Number of starting conditions to try}

\item{limit.slope}{("none", "on", "off")}
}
\description{
Fit Impulse Model (Single sigmoid)
}
\keyword{internal}
