% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse.start.single}
\alias{impulse.start.single}
\title{Starting Conditions for impulse model (single sigmoid)}
\usage{
impulse.start.single(x, y, k, limit.slope)
}
\description{
Function to determine k sets of initial starting conditions for impulse fitting
}
