% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-tips.R
\name{loadTipCells}
\alias{loadTipCells}
\title{Load Tip Cells}
\usage{
loadTipCells(object, tips)
}
\arguments{
\item{object}{An URD object}

\item{tips}{(Character) A clustering that was used as the tips (i.e. a column name of \code{@group.ids})}
}
\value{
An URD object with the cells that were used to start each tip stored in \code{@tree$cells.in.tip}
}
\description{
This loads the cells that belong to each tip into the URD object. This information
is required if \code{weighted.fusion=T} during \code{\link{buildTree}} so that
new segment visitation can properly be determined using a weighted average, based
on the number of cells that were in each tip.
}
\examples{
# Load the cells used for each tip into the URD object
axial.tree <- loadTipCells(axial, "tip.clusters")

}
