% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFDoublets.R
\name{NMFDoubletsDefineModules}
\alias{NMFDoubletsDefineModules}
\title{NMF Doublets: Calculate module pair overlap}
\usage{
NMFDoubletsDefineModules(object, modules.use, module.thresh.high = 0.3,
  module.thresh.low = 0.15)
}
\arguments{
\item{object}{An URD object}

\item{modules.use}{(Character vector) Modules to consider (i.e. column names of \code{@nmf.c1})}

\item{module.thresh.high}{(Numeric) High threshold to use for 'expression' of an NMF module}

\item{module.thresh.low}{(Numeric) Lower threshold to use for 'expression' of an NMF module}
}
\value{
(List) For use in \code{\link{NMFDoubletsPlotModuleThresholds}}
}
\description{
This considers NMF modules (as loaded into slot \code{@nmf.c1}) pairwise
to determine how many cells have overlapping expression of that pair of
modules. Two thresholds are used to determine whether a cell 'expresses'
a module -- this helps differentiate between totally distinct modules (in
which case the number of co-expressing cells is similar with either threshold)
from pairs of modules expressed in an overlapping gradient (in which case
the number of cells that co-express them increases dramatically as the threshold
is lowered).
}
