% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFDoublets.R
\name{NMFDoubletsPlotModuleCombos}
\alias{NMFDoubletsPlotModuleCombos}
\title{Plot NMF module combinations and dual-expressing cells}
\usage{
NMFDoubletsPlotModuleCombos(object, path, module.combos,
  module.expressed.thresh, frac.overlap.max, frac.overlap.diff.max,
  boundary = c("pass", "discarded"), sort = c("near", "random"),
  only.combos.with.doublets = T, n.plots = 50, width = 1400,
  height = 600)
}
\arguments{
\item{object}{An URD object}

\item{module.combos}{(List) Output from \code{\link{NMFDoubletsDefineModules}}}

\item{module.expressed.thresh}{(Numeric) Threshold for calling a cell as an expresser of a module}

\item{frac.overlap.max}{(Numeric) Maximum}

\item{frac.overlap.diff.max}{(Numeric) Max}

\item{boundary}{(Character) Should plots be for module combinations that \code{"pass"} the thresholds or that would be \code{"discarded"}}

\item{sort}{(Character) Should plots be for those module combinations \code{"near"} the boundary or chosen at \code{"random"}}

\item{only.combos.with.doublets}{(Logical) Should plots be limited to those module combinations that identify doublet cells?}

\item{n.plots}{(Numeric) What is the maximum number of plots that should be generated?}
}
\value{
A ggplot2 object
}
\description{
Plot NMF module combinations and dual-expressing cells
}
