% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dim.R
\name{plotDim}
\alias{plotDim}
\title{Dimensionality Reduction Plot}
\usage{
plotDim(
  object,
  label,
  label.type = "search",
  reduction.use = c("tsne", "pca", "dm"),
  dim.x = 1,
  dim.y = 2,
  colors = NULL,
  discrete.colors = NULL,
  point.size = 1,
  alpha = 1,
  point.shapes = F,
  plot.title = label,
  legend = T,
  legend.title = "",
  legend.point.size = 3 * point.size,
  label.clusters = F,
  cells = NULL,
  x.lim = NULL,
  y.lim = NULL,
  color.lim = NULL,
  na.rm = F,
  transitions.plot = 0,
  transitions.alpha = 0.5,
  transitions.df = NULL
)
}
\arguments{
\item{object}{An URD object}

\item{label}{(Character) Data to use for coloring points (e.g. a metadata name, group ID from clustering, or a gene name)}

\item{label.type}{(Character) Type of data to search for the label. Default is "search" which checks several data types in order. For more information: \code{\link{data.for.plot}}}

\item{reduction.use}{(Character) Dimensionality reduction to use (tSNE, PCA, or Diffusion Map)}

\item{dim.x}{(Numeric) Component to use on x-axis}

\item{dim.y}{(Numeric) Component to use on y-axis}

\item{colors}{(Character vector) Colors to use to generate a gradient scale to color continuous data}

\item{discrete.colors}{(Character vector) Colors to use to color}

\item{point.size}{(Numeric) Size of points on plot}

\item{alpha}{(Numeric) Transparency of points on plot: 0 (Transparent) - 1 (Opaque)}

\item{point.shapes}{(Logical) Should point shapes vary? This is useful in plots of discrete data types with many categories to help differentiate between similar colors.}

\item{plot.title}{(Character) Title of the plot}

\item{legend}{(Logical) Show a legend?}

\item{legend.title}{(Character) Should the legend be titled?}

\item{legend.point.size}{(Numeric) How big should points be in the legend?}

\item{label.clusters}{(Logical) Label centroids of a discrete label?}

\item{cells}{(Character vector) Cells to show on the plot (Default \code{NULL} is all cells.)}

\item{x.lim}{(Numeric) Limits of x-axis (NULL autodetects)}

\item{y.lim}{(Numeric) Limits of y-axis (NULL autodetects)}

\item{color.lim}{(Numeric) Limits of the point color scale (NULL autodetects)}

\item{na.rm}{(Logical) Should points with value NA for the desired data be removed from the plot?}

\item{transitions.plot}{(Numeric or NULL) Number of transition matrix connections to add to the plot. \code{NULL} will plot all connections. (WARNING: Too many connections will produce an unreadable plot that takes a long time to plot. Start with 10,000.)}

\item{transitions.alpha}{(Numeric) Maximum transparency of line segments representing transitions. (They are scaled based on their transition probability).}

\item{transitions.df}{(data.frame) Output from \link{edgesFromDM} (potentially further curated) to display on the plot. If provided, \code{transitions.plot} is ignored and all transitions in the provided data.frame are plotted.}
}
\value{
A ggplot2 object
}
\description{
Plots cells according to their coordinates in a dimensionality reduction (tSNE by default,
but also PCA or diffusion map). Cells are colored according to a user set \code{label} that
can range from gene expression, to metadata values, or cluster identity. See 
\code{\link{data.for.plot}} for more information about labels that can be chosen. Additionally, 
see \code{\link{plotDimDual}} to plot two
continuous variables simultaneously, \code{\link{plotDimHighlight}} to highlight one group
from a discrete label, \code{\link{plotDimArray}} to repeat the same plot across several
sets of dimensions, and \code{\link{plotDim3D}} to plot in three dimensions.. Additionally, 
\code{transitions.plot} can plot the connections from the diffusion map onto the plot.
}
