% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-smooth.R
\name{plotSmoothFit}
\alias{plotSmoothFit}
\title{Gene Smooth: Plot a single fit}
\usage{
plotSmoothFit(smoothed.fit, genes, scaled = T, multiplot = F,
  plot.data = T, alpha.data = 0.2, alpha.smooth = 1, lwd.smooth = 1,
  plot.title = NULL)
}
\arguments{
\item{smoothed.fit}{(List) Output from either \code{\link{geneSmoothFit}} or \code{\link{geneCascadeProcess}}}

\item{genes}{(Character vector) Genes to include in the plot}

\item{scaled}{(Logical) Plot actual expression values (\code{FALSE}) or expression scaled to its maximum value (\code{TRUE})}

\item{multiplot}{(Logical) If multiple genes are provided, they are plotted on the same plot in different colors if \code{multiplot=F} or as a grid of plots per gene if \code{multiplot=T}}

\item{plot.data}{(Logical) Plot data points?}

\item{alpha.data}{(Numeric: 0-1) Transparency of data points}

\item{alpha.smooth}{(Numeric: 0-1) Transparency of curve}

\item{lwd.smooth}{(Numeric) Line width of curve}

\item{plot.title}{(Character) Title of plot (Default \code{NULL} is no title)}
}
\value{
A ggplot2 object
}
\description{
This plots gene expression in groups of cells and a curve representing its mean expression (generated using a smoothing algorithm). It takes output from either \code{\link{geneSmoothFit}} or \code{\link{geneCascadeProcess}}. (Output from \code{\link{geneCascadeProcess}} may also be plotted with \code{\link{geneCascadeImpulsePlots}}.) If multiple genes are provided, they are plotted on the same plot in different colors if \code{multiplot=F} or as a grid of plots per gene if \code{multiplot=T}.
}
