% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{plotTreeForce}
\alias{plotTreeForce}
\title{Plot force-directed layout of tree}
\usage{
plotTreeForce(object, label, label.type = "search", view = "default",
  alpha = 0.8, alpha.fade = 0.1, size = 5, size.fade = 3,
  title = NULL, title.cex = 3, title.line = 0,
  label.tips = (!is.null(object@tree$segment.names) |
  !is.null(object@tree$segment.names.short)),
  use.short.names = !is.null(object@tree$segment.names.short),
  seg.show = NULL, cells.show = NULL, fade.below = (2/9),
  density.alpha = T, label.spacing = 5, text.cex = 0.8, colors = NULL,
  discrete.colors = NULL)
}
\arguments{
\item{object}{An URD object}

\item{label}{(Character)}

\item{label.type}{(Character)}

\item{view}{(Character) Can}

\item{alpha}{(Numeric) Maximum transparency of points}

\item{alpha.fade}{(Numeric) Minimum transparency of points}

\item{size}{(Numeric) Size of points in the plot}

\item{size.fade}{(Numeric) DOESN'T WORK.}

\item{title}{(Character) Title to add to the plot. Unfortunately, this provides rather ugly titles -- they are well positioned automatically for quick iteration, but not suitable for final figures.}

\item{title.cex}{(Numeric) Adjust the title font size}

\item{title.line}{(Numeric) Adjust the position of the title. Positive numbers move the title upward.}

\item{label.tips}{(Logical) Should text identifying the tips of the tree be placed in the force directed layout? Defaults to TRUE if \code{@tree$segment.names} has been set.}

\item{use.short.names}{(Logical) Should short names from \code{@tree$segment.names.short} be used? Defaults to TRUE if those values have been set.}

\item{seg.show}{(Character vector) Segments of the tree to put on the plot (Default \code{NULL} is all segments)}

\item{cells.show}{(Character vector) Cells of the tree to show (Default \code{NULL} is all cells)}

\item{fade.below}{(Numeric)}

\item{density.alpha}{(Logical)}

\item{label.spacing}{(Numeric)}

\item{text.cex}{(Numeric) Size of the label text}

\item{colors}{(Character vector)}

\item{discrete.colors}{(Character vector)}
}
\value{
Nothing. Produces a plot using the \code{rgl} package, displayed in an X11 window.
}
\description{
Plot force-directed layout of tree
}
