% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{processRandomWalks}
\alias{processRandomWalks}
\title{Process random walks and store in object}
\usage{
processRandomWalks(object, walks, walks.name, aggregate.fun = mean,
  n.subsample = 10, verbose = T)
}
\arguments{
\item{object}{An URD object}

\item{walks}{(List) List of character vectors of cells visited during random walks.}

\item{walks.name}{(Character) Name to use for storing walks}

\item{aggregate.fun}{(Function) Function to aggregate pseudotime (default: mean)}

\item{n.subsample}{(Numeric) Number of subsamplings to perform for calculating}

\item{verbose}{(Logical) Report on progress}
}
\value{
An URD object with cell visitation frequency stored in \code{@diff.data}, a calculated pseudotime stored in \code{@pseudotime}, and
subsampled data in \code{@pseudotime.stability}.
}
\description{
Process random walks and store in object
}
