% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{prop.exp}
\alias{prop.exp}
\title{Returns proportion of expressing cells}
\usage{
prop.exp(x)
}
\arguments{
\item{x}{(Numeric vector) Input values}
}
\description{
Determines proportion of input values > 0.
}
