% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-structure.R
\name{pseudotimeBreakpointByStretch}
\alias{pseudotimeBreakpointByStretch}
\title{Pseudotime Breakpoint By Stretch}
\usage{
pseudotimeBreakpointByStretch(
  div.pseudotime,
  segment.1,
  segment.2,
  visit.data,
  pseudotime.windows,
  verbose = T
)
}
\description{
Pseudotime Breakpoint By Stretch
}
\keyword{internal}
