% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{simulateRandomWalk}
\alias{simulateRandomWalk}
\title{Simulate random walks}
\usage{
simulateRandomWalk(start.cells, transition.matrix, end.cells, n = 10000,
  end.visits = 1, verbose.freq = 0, max.steps = 5000)
}
\arguments{
\item{start.cells}{(Character vector) Cells to use as a starting pool. One is chosen at random.}

\item{transition.matrix}{(Matrix or Sparse Matrix) Transition matrix}

\item{end.cells}{(Character vector)}

\item{n}{(Numeric) Number of walks to simulate}

\item{end.visits}{(Numeric)}

\item{verbose}{(Logical)}
}
\value{
(Character vector) Names of cells visited during the random walk.
}
\description{
Simulate random walks
}
