% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{simulateRandomWalksFromTips}
\alias{simulateRandomWalksFromTips}
\title{Simulate Random Walks From All Tips}
\usage{
simulateRandomWalksFromTips(object, tip.group.id, root.cells, transition.matrix,
  n.per.tip = 10000, root.visits = 1,
  max.steps = ncol(object@logupx.data), verbose = T)
}
\arguments{
\item{object}{An URD object}

\item{tip.group.id}{(Character) The name of the clustering that defines tips}

\item{root.cells}{(Character vector) Names of cells that constitute the root}

\item{transition.matrix}{(Matrix or dgCMatrix) Biased transition matrix}

\item{n.per.tip}{(Numeric) Number of walks to do per tip}

\item{root.visits}{(Numeric) Number of steps to take that visit a root.cell before stopping}

\item{verbose}{(Logical) Whether to report on progress}

\item{max.visits}{(Numeric) Abandon walks that take more steps than this, as it likely means that it has gotten stuck. (Default is number of cells in the data. On large data, may want to lower this value.)}
}
\value{
List (of tips) of lists (of walk paths) of character vectors
}
\description{
This automates
}
