% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-force-layout.R
\name{treeForceDirectedLayout}
\alias{treeForceDirectedLayout}
\title{Generate force-directed layout using tip-walk data.}
\usage{
treeForceDirectedLayout(object, num.nn = NULL, pseudotime, method = c("fr",
  "drl", "kk"), dim = 2, cells.to.do = NULL, cut.outlier.cells = NULL,
  cut.outlier.edges = NULL, cut.unconnected.segments = 1,
  min.final.neighbors = 2, tips = object@tree$tips,
  max.pseudotime.diff = NULL, plot.outlier.cuts = F, verbose = F,
  coords = "auto", start.temp = NULL, density.neighbors = 10)
}
\arguments{
\item{object}{An URD object}

\item{num.nn}{(Numeric) Number of nearest neighbors to use,}

\item{pseudotime}{(Character) Pseudotime to use (i.e. a column name of \code{@pseudotime}). (NULL omits pseudotime from the distance calculation of the nearest neighbor network.)}

\item{method}{(Character) Which force-directed layout algorithm to use}

\item{dim}{(Numeric) Don't change this from 2.}

\item{cells.to.do}{(Character vector) Cells to use in the layout (default \code{NULL} is all cells in the tree.)}

\item{cut.outlier.cells}{Who knows}

\item{cut.outlier.edges}{Who knows}

\item{cut.unconnected.segments}{(Numeric) Cut connections in the nearest-neighbor graph that are more
distant in the tree dendrogram structure than this. Default value is 1, which is the most aggressive.
It will only maintain connections within a segment and to that segment's parent or children. Higher values
permit more distant connections. (To disable, set to NA or NULL, which will permit any connections.)}

\item{min.final.neighbors}{(Numeric) After trimming outlier and unconnected connections in the nearest
neighbor graph,}
}
\description{
Generate force-directed layout using tip-walk data.
}
